Fix: refactor workspace context
