+++ frontend/src/context/WorkspaceContext.tsx
@@ -3,37 +3,16 @@
   useCallback,
   useContext,
   useEffect,
-  useRef,
-  useState,
 } from 'react';
 import {
-  Theme,
-  ThemePreference,
-  Locale,
-  Member,
-  Session,
-  Message,
-  BackendChatAgent,
-  BackendChatMessage,
   BackendChatSession,
-  BackendChatSessionAgent,
-  ChatActiveRun,
-  ChatSessionRuntimeSnapshot,
-  QuotedMessageReference,
-  Provider,
-  Strategy,
-  BackendChatSkill,
-  Config,
-  Environment,
-  MemberQueuesBySessionAgentId,
-  MemberQueueSnapshot,
-  QueuedMessageStatus,
-  UpdateChatSession,
-  WorkflowCardProjection,
-  WorkflowSessionStatusResponse,
-  WorkflowSidebarState,
-  WorkspaceChangesResponse,
   JsonValue,
+  Member,
+  MemberQueueSnapshot,
+  Message,
+  Provider,
+  QueuedMessageStatus,
+  WorkflowSessionStatusResponse,
 } from '@/types';
 import { i18nDict } from '@/i18n';
 import { mockFrontendApi } from '@/lib/mockFrontendApi';
@@ -55,1754 +34,75 @@ import {
 import {
   type CreateProjectRequest,
   type InboxItem,
-  type InboxSummary,
-  SoundFile,
-  type NotificationConfig,
-  type NotificationInboxSourcesConfig,
   type Project,
-  type ProjectMemberWithRuntime,
 } from '../../../shared/types';
 import {
   effectiveSessionAgentModelName,
-  mapMessage,
   mapMessages,
-  monogramFromName,
   mapProviders,
   mapSessionAgentsToMembers,
   mapSessions,
 } from '@/lib/mappers';
 import { resolveMessageReferences } from '@/lib/messageReferences';
-import {
-  AsyncResourceState,
-  beginLoad,
-  fail,
-  initialAsync,
-  succeed,
-} from '@/lib/asyncResource';
-import { notifyBuildStatsUsageUpdated } from '@/lib/buildStatsEvents';
-import { notifySourceControlRefreshRequested } from '@/lib/sourceControlEvents';
+import { beginLoad, fail, initialAsync, succeed } from '@/lib/asyncResource';
 import {
   hasRunningWorkflowActivity,
-  idleWorkflowSessionStatus,
   isWorkflowSidebarRunning,
-  resolveWorkflowSidebarState,
 } from '@/lib/workflowSidebarState';
-import { createConfigPatchQueue } from './configPatchQueue';
-import { useRunActivityStore } from './RunActivityContext';
-
-type ListUpdater<T> = T[] | ((prev: T[]) => T[]);
-
-type ChatInputMode = 'free' | 'workflow';
-const DEFAULT_CHAT_INPUT_MODE: ChatInputMode = 'free';
-type RuntimeActiveRun = ChatActiveRun;
-
-const resolveChatInputMode = (
-  value: string | null | undefined,
-): ChatInputMode => (value === 'workflow' ? 'workflow' : 'free');
-
-const toSessionChatInputMode = (mode: ChatInputMode): string | null =>
-  mode === 'workflow' ? 'workflow' : null;
-
-const chatSessionUpdatePayload = (
-  patch: Partial<UpdateChatSession>,
-): UpdateChatSession => ({
-  title: null,
-  status: null,
-  summary_text: null,
-  archive_ref: null,
-  last_seen_diff_key: null,
-  default_workspace_path: null,
-  ...patch,
-});
-
-interface SendMessageOptions {
-  chatInputMode?: ChatInputMode;
-  quotedMessage?: QuotedMessageReference;
-  routeMentions?: string[];
-  fallbackMention?: string | null;
-  workflowLeadAgentId?: string | null;
-  persistToBackend?: boolean;
-  placeholderMember?: Pick<Member, 'avatar' | 'name' | 'modelName'> | null;
-}
-
-export type ToastTone = 'info' | 'success' | 'warning' | 'error';
-
-export type WorkspaceToast = {
-  message: string;
-  tone: ToastTone;
-};
-
-export type WorkflowRuntimeLine = {
-  id: string;
-  executionId: string;
-  workflowAgentSessionId: string | null;
-  stepId: string;
-  stepKey: string;
-  agentId: string;
-  agentName: string;
-  streamType: 'assistant' | 'thinking' | 'error';
-  content: string;
-  createdAt: string;
-};
-
-type FileChangeType = 'created' | 'modified' | 'deleted';
-
-type FileChangeEntry = {
-  path: string;
-  change_type: FileChangeType;
-};
-
-type ChatStreamEvent =
-  | {
-      type: 'agent_run_started';
-      session_id: string;
-      session_agent_id: string;
-      agent_id: string;
-      agent_name: string;
-      model: string | null;
-      run_id: string;
-      source_message_id: string;
-      client_message_id: string | null;
-      started_at: string | null;
-    }
-  | {
-      type: 'agent_activity_updated';
-      session_id: string;
-      run_id: string;
-      latest_sequence: number;
-    }
-  | {
-      type: 'message_new' | 'message_updated';
-      message: BackendChatMessage;
-    }
-  | {
-      type: 'agent_state';
-      session_agent_id: string;
-      agent_id: string;
-      state: string;
-      run_id: string | null;
-      started_at: string | null;
-    }
-  | {
-      type: 'mention_error';
-      session_id: string;
-      message_id: string;
-      client_message_id: string | null;
-      agent_name: string;
-      agent_id: string | null;
-      reason: string;
-    }
-  | {
-      type: 'workflow_runtime_line';
-      line_id: string;
-      session_id: string;
-      execution_id: string;
-      workflow_agent_session_id: string | null;
-      step_id: string;
-      step_key: string;
-      agent_id: string;
-      agent_name: string;
-      stream_type: 'assistant' | 'thinking' | 'error';
-      content: string;
-      created_at: string;
-    }
-  | {
-      type: 'workflow_execution_updated';
-      session_id: string;
-      execution_id: string;
-    }
-  | {
-      type: 'file_change_refresh';
-      session_id: string;
-      session_agent_id: string;
-      agent_id: string;
-      run_id: string;
-      message_id: string;
-      changed_files: FileChangeEntry[];
-      ts: string;
-    }
-  | {
-      type: 'queue_updated';
-      session_id: string;
-      session_agent_id: string;
-      queue: MemberQueueSnapshot;
-    };
-
-const chatStreamWebSocketUrl = (path: string): string => {
-  const base =
-    typeof window === 'undefined' ? 'http://localhost' : window.location.href;
-  const url = new URL(path, base);
-  url.protocol = url.protocol === 'https:' ? 'wss:' : 'ws:';
-  return url.toString();
-};
-
-const PENDING_AGENT_MESSAGE_PREFIX = 'pending-agent-';
-const OPTIMISTIC_USER_MESSAGE_PREFIX = 'msg-user-';
-const QUEUE_PROCESSING_AGENT_MESSAGE_PREFIX =
-  `${PENDING_AGENT_MESSAGE_PREFIX}queue-processing-`;
-const createClientMessageId = (): string => {
-  const uuid = globalThis.crypto?.randomUUID?.();
-  return `${OPTIMISTIC_USER_MESSAGE_PREFIX}${
-    uuid ?? `${Date.now()}-${Math.random().toString(36).slice(2)}`
-  }`;
-};
-// Persist the user's last-viewed project/session so a page refresh restores the
-// same context (and therefore reconnects the WS stream to the same session)
-// instead of always falling back to the first session in the list.
-const ACTIVE_SESSION_ID_STORAGE_KEY = 'openteams-active-session-id';
-const SELECTED_PROJECT_ID_STORAGE_KEY = 'openteams-selected-project-id';
-const RUNNING_AGENT_SESSION_IDS_STORAGE_KEY =
-  'openteams-running-agent-session-ids';
-const UNREAD_AGENT_COMPLETION_SESSION_IDS_STORAGE_KEY =
-  'openteams-unread-agent-completion-session-ids';
-const ACKED_WORKFLOW_INPUT_IDS_STORAGE_KEY =
-  'openteams-acked-workflow-input-ids';
-const ACKED_WORKFLOW_ERROR_SESSION_IDS_STORAGE_KEY =
-  'openteams-acked-workflow-error-session-ids';
-// WebSocket auto-reconnect backoff bounds (ms).
-const CHAT_STREAM_RECONNECT_BASE_DELAY_MS = 1000;
-const CHAT_STREAM_RECONNECT_MAX_DELAY_MS = 30000;
-const SIDEBAR_RUNNING_INDICATOR_POLL_MS = 5000;
-const STARTING_AGENT_RECONCILE_DELAY_MS = 30000;
-const INBOX_REFRESH_INTERVAL_MS = 30000;
-const INBOX_LIGHT_REFRESH_DELAY_MS = 800;
-const INBOX_LIST_LIMIT = 25;
-const CHAT_MESSAGE_FONT_SIZE_DEFAULT = 14;
-export const CHAT_MESSAGE_FONT_SIZE_OPTIONS = [13, 14, 15, 16] as const;
-
-const EMPTY_INBOX_SUMMARY: InboxSummary = {
-  unread_count: 0n,
-  unread_by_severity: [],
-  unread_by_kind: [],
-};
-
-type InboxNotificationSource = keyof NotificationInboxSourcesConfig;
-
-const DEFAULT_NOTIFICATION_INBOX_SOURCES: NotificationInboxSourcesConfig = {
-  chat_message: true,
-  workflow_action: true,
-  approval: true,
-  worktree: true,
-  failure: true,
-};
-
-const DEFAULT_NOTIFICATION_CONFIG: NotificationConfig = {
-  sound_enabled: true,
-  push_enabled: true,
-  sound_file: SoundFile.ABSTRACT_SOUND3,
-  inbox_sources: DEFAULT_NOTIFICATION_INBOX_SOURCES,
-};
-
-const SOUND_FILE_VALUES = new Set<string>(Object.values(SoundFile));
-
-const notificationInboxSourcesFrom = (
-  value: unknown,
-): NotificationInboxSourcesConfig => {
-  const candidate =
-    value && typeof value === 'object' && !Array.isArray(value)
-      ? (value as Partial<Record<InboxNotificationSource, unknown>>)
-      : {};
-  return {
-    chat_message:
-      typeof candidate.chat_message === 'boolean'
-        ? candidate.chat_message
-        : DEFAULT_NOTIFICATION_INBOX_SOURCES.chat_message,
-    workflow_action:
-      typeof candidate.workflow_action === 'boolean'
-        ? candidate.workflow_action
-        : DEFAULT_NOTIFICATION_INBOX_SOURCES.workflow_action,
-    approval:
-      typeof candidate.approval === 'boolean'
-        ? candidate.approval
-        : DEFAULT_NOTIFICATION_INBOX_SOURCES.approval,
-    worktree:
-      typeof candidate.worktree === 'boolean'
-        ? candidate.worktree
-        : DEFAULT_NOTIFICATION_INBOX_SOURCES.worktree,
-    failure:
-      typeof candidate.failure === 'boolean'
-        ? candidate.failure
-        : DEFAULT_NOTIFICATION_INBOX_SOURCES.failure,
-  };
-};
-
-const notificationConfigFromWorkspaceConfig = (
-  config: Config | null,
-): NotificationConfig => {
-  const value = config?.notifications;
-  if (!value || typeof value !== 'object' || Array.isArray(value)) {
-    return DEFAULT_NOTIFICATION_CONFIG;
-  }
-
-  const candidate = value as {
-    sound_enabled?: unknown;
-    push_enabled?: unknown;
-    sound_file?: unknown;
-    inbox_sources?: unknown;
-  };
-  return {
-    sound_enabled:
-      typeof candidate.sound_enabled === 'boolean'
-        ? candidate.sound_enabled
-        : DEFAULT_NOTIFICATION_CONFIG.sound_enabled,
-    push_enabled:
-      typeof candidate.push_enabled === 'boolean'
-        ? candidate.push_enabled
-        : DEFAULT_NOTIFICATION_CONFIG.push_enabled,
-    sound_file:
-      typeof candidate.sound_file === 'string' &&
-      SOUND_FILE_VALUES.has(candidate.sound_file)
-        ? (candidate.sound_file as SoundFile)
-        : DEFAULT_NOTIFICATION_CONFIG.sound_file,
-    inbox_sources: notificationInboxSourcesFrom(candidate.inbox_sources),
-  };
-};
-
-const inboxNotificationSourceForKind = (
-  kind: string,
-): InboxNotificationSource | null => {
-  if (kind === 'chat_message') return 'chat_message';
-  if (
-    kind === 'chat_agent_failed' ||
-    kind === 'chat_mention_failed' ||
-    kind === 'workflow_execution_failed'
-  ) {
-    return 'failure';
-  }
-  if (kind === 'executor_approval') return 'approval';
-  if (kind.startsWith('worktree_')) return 'worktree';
-  if (kind.startsWith('workflow_')) return 'workflow_action';
-  return null;
-};
-
-const inboxSourceEnabledForKind = (
-  sources: NotificationInboxSourcesConfig,
-  kind: string,
-): boolean => {
-  const source = inboxNotificationSourceForKind(kind);
-  return source ? sources[source] : true;
-};
-
-const filterInboxItemsForEnabledSources = (
-  config: Config | null,
-  items: InboxItem[],
-): InboxItem[] => {
-  if (!config) return items;
-  const notificationConfig = notificationConfigFromWorkspaceConfig(config);
-  return items.filter((item) =>
-    inboxSourceEnabledForKind(notificationConfig.inbox_sources, item.kind),
-  );
-};
-
-const inboxCountValue = (value: bigint | number): bigint =>
-  typeof value === 'bigint' ? value : BigInt(value);
-
-const decrementInboxSummaryEntries = (
-  entries: InboxSummary['unread_by_kind'],
-  decrements: Map<string, number>,
-): InboxSummary['unread_by_kind'] =>
-  entries
-    .map((entry) => {
-      const decrement = decrements.get(entry.key) ?? 0;
-      if (decrement <= 0) return entry;
-      const nextCount = inboxCountValue(entry.count) - BigInt(decrement);
-      return {
-        ...entry,
-        count: nextCount > 0n ? nextCount : 0n,
-      };
-    })
-    .filter((entry) => inboxCountValue(entry.count) > 0n);
-
-const countUnreadInboxItems = (items: InboxItem[]) => {
-  const byKind = new Map<string, number>();
-  const bySeverity = new Map<string, number>();
-  let total = 0;
-  for (const item of items) {
-    if (item.read_at !== null || item.archived_at !== null) continue;
-    total += 1;
-    byKind.set(item.kind, (byKind.get(item.kind) ?? 0) + 1);
-    const severity = String(item.severity);
-    bySeverity.set(severity, (bySeverity.get(severity) ?? 0) + 1);
-  }
-  return { total, byKind, bySeverity };
-};
-
-const filterInboxSummaryForEnabledSources = (
-  config: Config | null,
-  summary: InboxSummary,
-): InboxSummary => {
-  if (!config) return summary;
-  const notificationConfig = notificationConfigFromWorkspaceConfig(config);
-  const unread_by_kind = summary.unread_by_kind.filter((entry) =>
-    inboxSourceEnabledForKind(notificationConfig.inbox_sources, entry.key),
-  );
-  const unread_count = unread_by_kind.reduce(
-    (total, entry) => total + inboxCountValue(entry.count),
-    0n,
-  );
-  return {
-    ...summary,
-    unread_count,
-    unread_by_kind,
-    unread_by_severity: [],
-  };
-};
-
-const inboxNotificationSettingsSignature = (config: Config | null): string => {
-  const notificationConfig = notificationConfigFromWorkspaceConfig(config);
-  return JSON.stringify({
-    push_enabled: notificationConfig.push_enabled,
-    sound_enabled: notificationConfig.sound_enabled,
-    sound_file: notificationConfig.sound_file,
-    inbox_sources: notificationConfig.inbox_sources,
-  });
-};
-
-const browserNotificationApi = (): typeof Notification | null => {
-  if (typeof window === 'undefined' || !('Notification' in window)) {
-    return null;
-  }
-  return window.Notification;
-};
-
-const showInboxSystemNotification = (
-  config: Config | null,
-  newItems: InboxItem[],
-) => {
-  if (!config || newItems.length === 0) return;
-
-  const notificationConfig = notificationConfigFromWorkspaceConfig(config);
-  if (!notificationConfig.push_enabled) return;
-
-  const BrowserNotification = browserNotificationApi();
-  if (!BrowserNotification || BrowserNotification.permission !== 'granted') {
-    return;
-  }
-
-  newItems.forEach((item) => {
-    const notification = new BrowserNotification(
-      item.title || 'OpenTeams notification',
-      {
-        body: item.body?.trim() || undefined,
-        tag: `openteams-inbox-${item.id}`,
-        silent: true,
-      },
-    );
-    notification.onclick = () => {
-      window.focus();
-      notification.close();
-    };
-  });
-};
-
-const playInboxNotificationSound = (
-  config: Config | null,
-  newItems: InboxItem[],
-) => {
-  if (!config) return;
-  if (typeof Audio === 'undefined' || newItems.length === 0) return;
-
-  const notificationConfig = notificationConfigFromWorkspaceConfig(config);
-  if (!notificationConfig.sound_enabled) return;
-
-  const audio = new Audio(
-    `/api/sounds/${encodeURIComponent(notificationConfig.sound_file)}`,
-  );
-  void audio.play().catch(() => undefined);
-};
-
-const isAutoReadableInboxItem = (item: InboxItem): boolean =>
-  item.kind === 'chat_message' || item.source_type === 'chat_message';
-
-const isThemePreference = (
-  value: string | null | undefined,
-): value is ThemePreference =>
-  value === 'light' || value === 'dark' || value === 'system';
-
-const resolveSystemTheme = (): Theme => {
-  if (typeof window === 'undefined') return 'dark';
-  return window.matchMedia?.('(prefers-color-scheme: light)').matches
-    ? 'light'
-    : 'dark';
-};
-
-const readSessionIdSet = (storageKey: string): Set<string> => {
-  if (typeof localStorage === 'undefined') return new Set();
-  try {
-    const raw = localStorage.getItem(storageKey);
-    const parsed = raw ? JSON.parse(raw) : [];
-    if (!Array.isArray(parsed)) return new Set();
-    return new Set(
-      parsed.filter(
-        (value): value is string =>
-          typeof value === 'string' && value.trim().length > 0,
-      ),
-    );
-  } catch {
-    return new Set();
-  }
-};
-
-const writeSessionIdSet = (storageKey: string, sessionIds: Set<string>) => {
-  if (typeof localStorage === 'undefined') return;
-  try {
-    if (sessionIds.size === 0) {
-      localStorage.removeItem(storageKey);
-      return;
-    }
-    localStorage.setItem(storageKey, JSON.stringify([...sessionIds]));
-  } catch {}
-};
-
-const normalizeChatMessageFontSize = (value: number | string | null): number => {
-  const numeric = typeof value === 'number' ? value : Number(value);
-  if (!Number.isFinite(numeric)) return CHAT_MESSAGE_FONT_SIZE_DEFAULT;
-
-  const rounded = Math.round(numeric);
-  return (
-    CHAT_MESSAGE_FONT_SIZE_OPTIONS.find((option) => option === rounded) ??
-    CHAT_MESSAGE_FONT_SIZE_DEFAULT
-  );
-};
-
-const themePreferenceFromConfig = (theme: Config['theme']): ThemePreference => {
-  const normalized = theme.toLowerCase();
-  return isThemePreference(normalized) ? normalized : 'dark';
-};
-
-const themePreferenceToConfig = (
-  theme: ThemePreference,
-): Config['theme'] => theme.toUpperCase() as Config['theme'];
-
-const resolveBrowserLocale = (): Locale => {
-  if (typeof navigator === 'undefined') return 'zh';
-  const language = navigator.language.toLowerCase();
-  if (language.startsWith('en')) return 'en';
-  if (language.startsWith('ja')) return 'ja';
-  if (language.startsWith('ko')) return 'ko';
-  if (language.startsWith('fr')) return 'fr';
-  if (language.startsWith('es')) return 'es';
-  return 'zh';
-};
-
-const localeFromConfig = (language: Config['language']): Locale => {
-  switch (language) {
-    case 'EN':
-      return 'en';
-    case 'JA':
-      return 'ja';
-    case 'KO':
-      return 'ko';
-    case 'FR':
-      return 'fr';
-    case 'ES':
-      return 'es';
-    case 'ZH_HANS':
-    case 'ZH_HANT':
-      return 'zh';
-    case 'BROWSER':
-    default:
-      return resolveBrowserLocale();
-  }
-};
-
-const localeToConfig = (locale: Locale): Config['language'] => {
-  switch (locale) {
-    case 'en':
-      return 'EN';
-    case 'ja':
-      return 'JA';
-    case 'ko':
-      return 'KO';
-    case 'fr':
-      return 'FR';
-    case 'es':
-      return 'ES';
-    case 'zh':
-    default:
-      return 'ZH_HANS';
-  }
-};
-
-const chatMessageFontSizeFromConfig = (
-  value: Config['chat_bubble_font_size'],
-): number => normalizeChatMessageFontSize(value.replace(/^px/, ''));
-
-const chatMessageFontSizeToConfig = (
-  value: number,
-): Config['chat_bubble_font_size'] =>
-  `px${normalizeChatMessageFontSize(value)}` as Config['chat_bubble_font_size'];
-
-const isPendingAgentPlaceholder = (message: Message): boolean =>
-  Boolean(
-    message.isAgentRunning &&
-    !message.runId &&
-    message.id.startsWith(PENDING_AGENT_MESSAGE_PREFIX),
-  );
-
-const isActiveAgentState = (state: string | undefined): boolean =>
-  state === 'running' || state === 'stopping' || state === 'waitingapproval';
-
-const isRunningSessionAgentState = (state: string | undefined): boolean =>
-  state === 'running' || state === 'stopping';
-
-const hasRunningSessionAgent = (
-  sessionAgents: BackendChatSessionAgent[],
-  ignoredSessionAgentIds?: ReadonlySet<string>,
-): boolean =>
-  sessionAgents.some((sessionAgent) =>
-    !ignoredSessionAgentIds?.has(sessionAgent.id) &&
-    isRunningSessionAgentState(sessionAgent.state),
-  );
-
-type SessionRunningIndicators = {
-  hasRunningAgent: boolean;
-  hasRunningWorkflow: boolean;
-  workflowSidebarState: WorkflowSidebarState;
-  pendingWorkflowInputId: string | null;
-  pendingWorkflowReviewId: string | null;
-};
-
-const loadSessionRunningIndicators = async (
-  sessionIds: string[],
-  ignoredSessionAgentIds?: ReadonlySet<string>,
-  options?: { skipAgentSessionIds?: ReadonlySet<string> },
-): Promise<Map<string, SessionRunningIndicators>> => {
-  const entries = await Promise.all(
-    sessionIds.map(async (sessionId) => {
-      const shouldSkipAgents =
-        options?.skipAgentSessionIds?.has(sessionId) ?? false;
-      const [sessionAgents, workflowStatus] = await Promise.all([
-        shouldSkipAgents
-          ? Promise.resolve<BackendChatSessionAgent[]>([])
-          : sessionAgentsApi.list(sessionId).catch(() => []),
-        workflowApi
-          .getSessionStatus(sessionId)
-          .catch(() => idleWorkflowSessionStatus),
-      ]);
-      const workflowSidebarState = resolveWorkflowSidebarState(workflowStatus);
-      return [
-        sessionId,
-        {
-          hasRunningAgent: hasRunningSessionAgent(
-            sessionAgents,
-            ignoredSessionAgentIds,
-          ),
-          hasRunningWorkflow: isWorkflowSidebarRunning(workflowSidebarState),
-          workflowSidebarState,
-          pendingWorkflowInputId:
-            workflowStatus.pending_workflow_input_id ?? null,
-          pendingWorkflowReviewId:
-            workflowStatus.pending_workflow_review_id ?? null,
-        },
-      ] as const;
-    }),
-  );
-
-  return new Map(entries);
-};
-
-const isOptimisticUserMessage = (message: Message): boolean =>
-  Boolean(
-    message.isUser &&
-    message.id.startsWith(OPTIMISTIC_USER_MESSAGE_PREFIX),
-  );
-
-const isOptimisticPendingAgentPlaceholder = (message: Message): boolean =>
-  isPendingAgentPlaceholder(message) &&
-  (message.id.startsWith(
-    `${PENDING_AGENT_MESSAGE_PREFIX}${OPTIMISTIC_USER_MESSAGE_PREFIX}`,
-  ) || message.id.startsWith(QUEUE_PROCESSING_AGENT_MESSAGE_PREFIX));
-
-const userMessageClientId = (message: Message): string | undefined =>
-  message.clientMessageId ??
-  (isOptimisticUserMessage(message) ? message.id : undefined);
-
-const messageIdentityKeys = (message: Message): string[] => {
-  const keys = new Set<string>();
-  if (message.id) keys.add(message.id);
-  const clientMessageId = userMessageClientId(message);
-  if (clientMessageId) keys.add(clientMessageId);
-  return [...keys];
-};
-
-const firstMessageSourceKey = (
-  message: Message,
-  sourceKeys: Set<string>,
-): string | null => {
-  if (message.isUser) return null;
-  if (message.sourceMessageId && sourceKeys.has(message.sourceMessageId)) {
-    return message.sourceMessageId;
-  }
-  if (message.clientMessageId && sourceKeys.has(message.clientMessageId)) {
-    return message.clientMessageId;
-  }
-  return null;
-};
-
-const orderMessagesForConversation = (messages: Message[]): Message[] => {
-  const sourceKeys = new Set<string>();
-  for (const message of messages) {
-    if (!message.isUser) continue;
-    for (const key of messageIdentityKeys(message)) {
-      sourceKeys.add(key);
-    }
-  }
-
-  if (sourceKeys.size === 0) return messages;
-
-  const anchoredMessages = new Set<Message>();
-  const anchoredBySourceKey = new Map<string, Message[]>();
-  for (const message of messages) {
-    const sourceKey = firstMessageSourceKey(message, sourceKeys);
-    if (!sourceKey) continue;
-    anchoredMessages.add(message);
-    const anchored = anchoredBySourceKey.get(sourceKey) ?? [];
-    anchored.push(message);
-    anchoredBySourceKey.set(sourceKey, anchored);
-  }
-
-  if (anchoredMessages.size === 0) return messages;
-
-  const emittedAnchored = new Set<Message>();
-  const ordered: Message[] = [];
-  for (const message of messages) {
-    if (anchoredMessages.has(message)) continue;
-
-    ordered.push(message);
-    if (!message.isUser) continue;
-
-    for (const key of messageIdentityKeys(message)) {
-      const anchored = anchoredBySourceKey.get(key);
-      if (!anchored) continue;
-      for (const anchoredMessage of anchored) {
-        if (emittedAnchored.has(anchoredMessage)) continue;
-        ordered.push(anchoredMessage);
-        emittedAnchored.add(anchoredMessage);
-      }
-    }
-  }
-
-  for (const message of messages) {
-    if (anchoredMessages.has(message) && !emittedAnchored.has(message)) {
-      ordered.push(message);
-    }
-  }
-
-  return ordered;
-};
-
-const messageCreatedAtMs = (message: Message): number | null => {
-  if (!message.createdAt) return null;
-  const value = Date.parse(message.createdAt);
-  return Number.isNaN(value) ? null : value;
-};
-
-const insertMessageByCreatedAt = (
-  messages: Message[],
-  message: Message,
-): Message[] => {
-  const messageAt = messageCreatedAtMs(message);
-  if (messageAt === null) return [...messages, message];
-
-  const next = [...messages];
-  const index = next.findIndex((candidate) => {
-    const candidateAt = messageCreatedAtMs(candidate);
-    return candidateAt !== null && candidateAt > messageAt;
-  });
-  next.splice(index >= 0 ? index : next.length, 0, message);
-  return next;
-};
-
-const matchesUserMessageIdentity = (
-  message: Message,
-  messageId: string,
-  clientMessageId?: string,
-): boolean =>
-  Boolean(
-    message.isUser &&
-      (message.id === messageId ||
-        (clientMessageId && userMessageClientId(message) === clientMessageId)),
-  );
-
-const queuedChatMessageKeysForSession = (
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): Set<string> => {
-  const keys = new Set<string>();
-  for (const queue of queues) {
-    if (queue.session_id !== sessionId) continue;
-    for (const item of queue.items) {
-      if (String(item.message.status) !== 'queued') continue;
-      keys.add(item.message.chat_message_id);
-    }
-  }
-  return keys;
-};
-
-const queuedSessionAgentIdsByMessageKey = (
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): Map<string, Set<string>> => {
-  const targets = new Map<string, Set<string>>();
-  for (const queue of queues) {
-    if (queue.session_id !== sessionId) continue;
-    for (const item of queue.items) {
-      if (String(item.message.status) !== 'queued') continue;
-      const key = item.message.chat_message_id;
-      const sessionAgentIds = targets.get(key) ?? new Set<string>();
-      sessionAgentIds.add(queue.session_agent_id);
-      targets.set(key, sessionAgentIds);
-    }
-  }
-  return targets;
-};
-
-export const isStartingPlaceholderRepresentedInQueue = (
-  message: Message,
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): boolean => {
-  if (!isPendingAgentPlaceholder(message) || !message.sessionAgentId) {
-    return false;
-  }
-  const sessionAgentId = message.sessionAgentId;
-  const queuedTargets = queuedSessionAgentIdsByMessageKey(queues, sessionId);
-  const messageKeys = [message.sourceMessageId, message.clientMessageId].filter(
-    (key): key is string => Boolean(key),
-  );
-  return messageKeys.some((key) =>
-    queuedTargets.get(key)?.has(sessionAgentId),
-  );
-};
-
-const isQueuedUserMessageFromSnapshot = (
-  message: Message,
-  queuedMessageKeys: ReadonlySet<string>,
-): boolean => {
-  if (!message.isUser || queuedMessageKeys.size === 0) return false;
-  const clientMessageId = userMessageClientId(message);
-  return (
-    queuedMessageKeys.has(message.id) ||
-    Boolean(clientMessageId && queuedMessageKeys.has(clientMessageId))
-  );
-};
-
-const isQueuedPendingPlaceholderFromSnapshot = (
-  message: Message,
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): boolean =>
-  isStartingPlaceholderRepresentedInQueue(message, queues, sessionId);
-
-const filterQueuedUserMessagesFromSnapshot = (
-  messages: Message[],
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): Message[] => {
-  const queuedMessageKeys = queuedChatMessageKeysForSession(queues, sessionId);
-  if (queuedMessageKeys.size === 0) return messages;
-  const queuedTargetsByMessageKey = queuedSessionAgentIdsByMessageKey(
-    queues,
-    sessionId,
-  );
-  return messages.filter(
-    (message) => {
-      if (
-        isQueuedPendingPlaceholderFromSnapshot(
-          message,
-          queues,
-          sessionId,
-        )
-      ) {
-        return false;
-      }
-      if (!isQueuedUserMessageFromSnapshot(message, queuedMessageKeys)) {
-        return true;
-      }
-
-      const clientMessageId = userMessageClientId(message);
-      const messageKey = queuedMessageKeys.has(message.id)
-        ? message.id
-        : clientMessageId;
-      const queuedTargets = messageKey
-        ? queuedTargetsByMessageKey.get(messageKey)
-        : undefined;
-      return messages.some(
-        (candidate) =>
-          candidate.isAgentRunning &&
-          candidate.clientMessageId === clientMessageId &&
-          (!candidate.sessionAgentId ||
-            !queuedTargets?.has(candidate.sessionAgentId)),
-      );
-    },
-  );
-};
-
-const queuedUserMessagesByIdFromSnapshot = (
-  messages: Message[],
-  queues: ReadonlyArray<MemberQueueSnapshot>,
-  sessionId: string,
-): Record<string, Message> => {
-  const queuedMessageKeys = queuedChatMessageKeysForSession(queues, sessionId);
-  if (queuedMessageKeys.size === 0) return {};
-
-  const result: Record<string, Message> = {};
-  for (const message of messages) {
-    if (!isQueuedUserMessageFromSnapshot(message, queuedMessageKeys)) continue;
-    result[message.id] = message;
-    const clientMessageId = userMessageClientId(message);
-    if (clientMessageId) {
-      result[clientMessageId] = message;
-    }
-  }
-  return result;
-};
-
-type PendingPlaceholderMatch = {
-  sessionAgentId?: string;
-  clientMessageId?: string | null;
-  sourceMessageId?: string | null;
-  agentName?: string | null;
-};
-
-const normalizedAgentHandle = (name: string): string =>
-  name.replace(/^@/, '').trim().toLocaleLowerCase();
-
-const normalizePendingPlaceholderMatch = (
-  match?: string | PendingPlaceholderMatch,
-): PendingPlaceholderMatch => {
-  if (typeof match === 'string') return { sessionAgentId: match };
-  return match ?? {};
-};
-
-const pendingPlaceholderMatches = (
-  message: Message,
-  match: PendingPlaceholderMatch,
-): boolean => {
-  if (!isPendingAgentPlaceholder(message)) return false;
-  const hasCorrelationId = Boolean(
-    match.clientMessageId || match.sourceMessageId,
-  );
-  const correlationMatches = match.clientMessageId
-    ? message.clientMessageId === match.clientMessageId
-    : match.sourceMessageId
-      ? message.sourceMessageId === match.sourceMessageId
-      : true;
-  if (hasCorrelationId && !correlationMatches) return false;
-
-  const sessionAgentMatches = Boolean(
-    match.sessionAgentId &&
-      message.sessionAgentId === match.sessionAgentId,
-  );
-  const agentNameMatches = Boolean(
-    match.agentName &&
-      normalizedAgentHandle(message.sender) ===
-        normalizedAgentHandle(match.agentName),
-  );
-  if (match.sessionAgentId || match.agentName) {
-    return sessionAgentMatches || agentNameMatches;
-  }
-  return hasCorrelationId;
-};
-
-const findPendingAgentPlaceholderIndex = (
-  messages: Message[],
-  match?: string | PendingPlaceholderMatch,
-): number => {
-  const normalized = normalizePendingPlaceholderMatch(match);
-  if (
-    normalized.clientMessageId ||
-    normalized.sourceMessageId ||
-    normalized.sessionAgentId ||
-    normalized.agentName
-  ) {
-    for (let index = messages.length - 1; index >= 0; index -= 1) {
-      if (pendingPlaceholderMatches(messages[index], normalized)) {
-        return index;
-      }
-    }
-    return -1;
-  }
-
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    if (isPendingAgentPlaceholder(messages[index])) {
-      return index;
-    }
-  }
-
-  return -1;
-};
-
-// A session agent runs at most one run at a time. When a new run starts (or
-// its first activity line arrives), drop any prior running placeholder for a
-// *different* run of the same agent so a stale one 鈥?e.g. left over from a
-// just-stopped run that refreshMessages re-hydrated 鈥?cannot coexist with the
-// new run and produce duplicate "executing" placeholders.
-const evictStaleRunPlaceholders = (
-  messages: Message[],
-  sessionAgentId: string | undefined,
-  runId: string,
-): Message[] => {
-  if (!sessionAgentId) return messages;
-  return messages.filter(
-    (message) =>
-      !(
-        message.isAgentRunning &&
-        message.sessionAgentId === sessionAgentId &&
-        Boolean(message.runId) &&
-        message.runId !== runId
-      ),
-  );
-};
-
-export const mergeCarriedRunPlaceholder = (
-  existing: Message | undefined,
-  incoming: Message,
-): Message => {
-  if (!existing) return incoming;
-  if (existing.runId) return existing;
-
-  const primary = incoming;
-  const secondary = existing;
-  const sourceMessageId =
-    primary.sourceMessageId ?? secondary.sourceMessageId;
-  const clientMessageId =
-    primary.clientMessageId ?? secondary.clientMessageId;
-  const secondaryHasAnchor = Boolean(
-    secondary.sourceMessageId || secondary.clientMessageId,
-  );
-
-  return {
-    ...primary,
-    id: existing.id,
-    sourceMessageId,
-    clientMessageId,
-    createdAt:
-      secondaryHasAnchor && (sourceMessageId || clientMessageId)
-        ? (secondary.createdAt ?? primary.createdAt)
-        : (primary.createdAt ?? secondary.createdAt),
-  };
-};
-
-export const correlateRunningPlaceholdersWithPending = (
-  current: Message[],
-  runningPlaceholders: Message[],
-): { current: Message[]; runningPlaceholders: Message[] } => {
-  if (runningPlaceholders.length === 0) {
-    return { current, runningPlaceholders };
-  }
-
-  const nextCurrent = [...current];
-  const unmatchedRunningPlaceholders: Message[] = [];
-  for (const running of runningPlaceholders) {
-    let pendingIndex = findPendingAgentPlaceholderIndex(nextCurrent, {
-      sessionAgentId: running.sessionAgentId,
-      clientMessageId: running.clientMessageId,
-      sourceMessageId: running.sourceMessageId,
-      agentName: running.sender,
-    });
-    if (pendingIndex < 0 && running.clientMessageId) {
-      const sameMessageCandidates = nextCurrent
-        .map((message, index) => ({ message, index }))
-        .filter(
-          ({ message }) =>
-            isPendingAgentPlaceholder(message) &&
-            message.clientMessageId === running.clientMessageId,
-        );
-      if (sameMessageCandidates.length === 1) {
-        pendingIndex = sameMessageCandidates[0].index;
-      }
-    }
-    if (pendingIndex < 0) {
-      unmatchedRunningPlaceholders.push(running);
-      continue;
-    }
-
-    const pending = nextCurrent[pendingIndex];
-    const upgraded: Message = {
-      ...running,
-      id: pending.id,
-      sourceMessageId: pending.sourceMessageId ?? running.sourceMessageId,
-      clientMessageId: pending.clientMessageId ?? running.clientMessageId,
-      createdAt: pending.createdAt ?? running.createdAt,
-    };
-    nextCurrent[pendingIndex] = upgraded;
-  }
-
-  return {
-    current: nextCurrent,
-    runningPlaceholders: unmatchedRunningPlaceholders,
-  };
-};
-
-export const findRunningPlaceholderIndexesForIncoming = (
-  current: Message[],
-  incoming: Message,
-): number[] => {
-  if (incoming.isUser) return [];
-
-  const candidates = current
-    .map((message, index) => ({ message, index }))
-    .filter(({ message }) => message.isAgentRunning);
-  const correlationCompatible = (message: Message): boolean => {
-    if (
-      incoming.clientMessageId &&
-      message.clientMessageId &&
-      incoming.clientMessageId !== message.clientMessageId
-    ) {
-      return false;
-    }
-    if (
-      incoming.sourceMessageId &&
-      message.sourceMessageId &&
-      incoming.sourceMessageId !== message.sourceMessageId
-    ) {
-      return false;
-    }
-    return true;
-  };
-  const compatible = candidates.filter(({ message }) =>
-    correlationCompatible(message),
-  );
-
-  if (incoming.runId) {
-    const runMatches = compatible.filter(
-      ({ message }) => message.runId === incoming.runId,
-    );
-    if (runMatches.length > 0) return runMatches.map(({ index }) => index);
-  }
-  if (incoming.sessionAgentId) {
-    const sessionAgentMatches = compatible.filter(
-      ({ message }) =>
-        message.sessionAgentId === incoming.sessionAgentId,
-    );
-    if (sessionAgentMatches.length > 0) {
-      return sessionAgentMatches.map(({ index }) => index);
-    }
-  }
-
-  const correlated = compatible.filter(({ message }) =>
-    incoming.clientMessageId
-      ? message.clientMessageId === incoming.clientMessageId
-      : incoming.sourceMessageId
-        ? message.sourceMessageId === incoming.sourceMessageId
-        : false,
-  );
-  const senderMatches = correlated.filter(
-    ({ message }) =>
-      normalizedAgentHandle(message.sender) ===
-      normalizedAgentHandle(incoming.sender),
-  );
-  if (senderMatches.length > 0) return senderMatches.map(({ index }) => index);
-  return correlated.length === 1 ? [correlated[0].index] : [];
-};
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-  typeof value === 'object' && value !== null && !Array.isArray(value);
-
-const hasNonNegativeNumberField = (
-  value: Record<string, unknown>,
-  fieldNames: string[],
-): boolean =>
-  fieldNames.some((fieldName) => {
-    const raw = value[fieldName];
-    return typeof raw === 'number' && Number.isFinite(raw) && raw >= 0;
-  });
-
-const hasCompleteTokenUsageBreakdown = (value: unknown): boolean => {
-  if (!isRecord(value)) return false;
-  return (
-    hasNonNegativeNumberField(value, ['input_tokens', 'snapshot_input_tokens']) &&
-    hasNonNegativeNumberField(value, [
-      'output_tokens',
-      'snapshot_output_tokens',
-    ])
-  );
-};
-
-const hasRealCompleteTokenUsage = (message: BackendChatMessage): boolean => {
-  if (message.sender_type !== 'agent' || !isRecord(message.meta)) return false;
-  const tokenUsage = message.meta.token_usage;
-  if (!isRecord(tokenUsage)) return false;
-  if (tokenUsage.is_estimated === true) return false;
-  return (
-    hasCompleteTokenUsageBreakdown(tokenUsage) ||
-    hasCompleteTokenUsageBreakdown(tokenUsage.last_token_usage) ||
-    hasCompleteTokenUsageBreakdown(tokenUsage.total_token_usage)
-  );
-};
-
-const firstNumberField = (
-  value: Record<string, unknown>,
-  fieldNames: string[],
-): number | null => {
-  for (const fieldName of fieldNames) {
-    const raw = value[fieldName];
-    if (typeof raw === 'number' && Number.isFinite(raw)) return raw;
-  }
-  return null;
-};
-
-const tokenUsageBreakdownSignature = (value: unknown) => {
-  if (!isRecord(value)) return null;
-  return {
-    input: firstNumberField(value, ['input_tokens', 'snapshot_input_tokens']),
-    output: firstNumberField(value, [
-      'output_tokens',
-      'snapshot_output_tokens',
-    ]),
-    cacheRead: firstNumberField(value, [
-      'cache_read_tokens',
-      'snapshot_cache_read_tokens',
-    ]),
-    reasoningOutput: firstNumberField(value, [
-      'reasoning_output_tokens',
-      'snapshot_reasoning_output_tokens',
-    ]),
-    total: firstNumberField(value, ['total_tokens', 'snapshot_total_tokens']),
-  };
-};
-
-const tokenUsageNotificationSignature = (
-  message: BackendChatMessage,
-): string | null => {
-  if (!hasRealCompleteTokenUsage(message) || !isRecord(message.meta)) {
-    return null;
-  }
-  const tokenUsage = message.meta.token_usage;
-  if (!isRecord(tokenUsage)) return null;
-
-  return JSON.stringify({
-    direct: tokenUsageBreakdownSignature(tokenUsage),
-    last: tokenUsageBreakdownSignature(tokenUsage.last_token_usage),
-    total: tokenUsageBreakdownSignature(tokenUsage.total_token_usage),
-  });
-};
-
-const extractAgentMentions = (text: string): string[] =>
-  Array.from(text.matchAll(/@([\p{L}\p{N}_-]+)/gu), (match) =>
-    match[1],
-  );
-
-const asAgentHandle = (name: string): string =>
-  name.startsWith('@') ? name : `@${name}`;
-
-export const memberNotFoundToastMessage = (
-  locale: Locale,
-  memberName: string,
-): string => {
-  const key = 'toast.memberNotFound';
-  const template = i18nDict[locale]?.[key] ?? i18nDict.en[key] ?? key;
-  return template.replace('{member}', asAgentHandle(memberName));
-};
-
-const optimisticAgentPlaceholderId = (
-  clientMessageId: string,
-  sessionAgentId: string | undefined,
-  sender: string,
-): string => {
-  const targetKey = sessionAgentId ?? normalizedAgentHandle(sender);
-  return `${PENDING_AGENT_MESSAGE_PREFIX}${clientMessageId}-${encodeURIComponent(targetKey)}`;
-};
-
-const makePendingAgentPlaceholder = (
-  targetMention: string,
-  userMsgId: string,
-  members: Member[],
-  sessionId: string,
-  placeholderMember?: Pick<Member, 'avatar' | 'name' | 'modelName'> | null,
-): Message | null => {
-  const normalizedMention = normalizedAgentHandle(targetMention);
-  if (!normalizedMention) return null;
-
-  const fallbackMember = members.find(
-    (member) => normalizedAgentHandle(member.name) === normalizedMention,
-  );
-  const matchingPlaceholderMember =
-    placeholderMember &&
-    normalizedAgentHandle(placeholderMember.name) === normalizedMention
-      ? placeholderMember
-      : null;
-  const displayMember = fallbackMember ?? matchingPlaceholderMember ?? null;
-  const sender = asAgentHandle(displayMember?.name ?? targetMention);
-
-  return {
-    id: optimisticAgentPlaceholderId(userMsgId, fallbackMember?.id, sender),
-    sessionId,
-    avatar: displayMember?.avatar ?? monogramFromName(sender),
-    sender,
-    model: displayMember?.modelName,
-    time: 'just now',
-    createdAt: new Date().toISOString(),
-    text: '',
-    isAgent: true,
-    isThinking: true,
-    isAgentRunning: true,
-    clientMessageId: userMsgId,
-    sessionAgentId: fallbackMember?.id,
-  };
-};
-
-export const makePendingAgentPlaceholders = (
-  targetMentions: string[],
-  userMsgId: string,
-  members: Member[],
-  sessionId: string,
-  placeholderMember?: Pick<Member, 'avatar' | 'name' | 'modelName'> | null,
-): Message[] => {
-  const uniqueTargets = new Map<string, string>();
-  for (const mention of targetMentions) {
-    const key = normalizedAgentHandle(mention);
-    if (key && !uniqueTargets.has(key)) uniqueTargets.set(key, mention);
-  }
-
-  return [...uniqueTargets.values()]
-    .map((mention) =>
-      makePendingAgentPlaceholder(
-        mention,
-        userMsgId,
-        members,
-        sessionId,
-        placeholderMember,
-      ),
-    )
-    .filter((message): message is Message => message !== null);
-};
-
-const queueProcessingPlaceholderId = (
-  chatMessageId: string,
-  sessionAgentId: string,
-): string =>
-  `${QUEUE_PROCESSING_AGENT_MESSAGE_PREFIX}${chatMessageId}-${sessionAgentId}`;
-
-const isQueueProcessingPlaceholder = (message: Message): boolean =>
-  message.id.startsWith(QUEUE_PROCESSING_AGENT_MESSAGE_PREFIX);
-
-export const reconcileProcessingQueuePlaceholders = (
-  current: Message[],
-  queue: MemberQueueSnapshot,
-  members: Member[],
-): Message[] => {
-  const activeQueueItems = queue.items.filter((item) =>
-    ['processing', 'running'].includes(String(item.message.status)),
-  );
-  const activeSourceMessageIds = new Set(
-    activeQueueItems.map((item) => item.message.chat_message_id),
-  );
-  const next = current.filter((message) => {
-    if (
-      !isQueueProcessingPlaceholder(message) ||
-      message.sessionAgentId !== queue.session_agent_id ||
-      message.runId
-    ) {
-      return true;
-    }
-    return Boolean(
-      message.sourceMessageId &&
-      activeSourceMessageIds.has(message.sourceMessageId),
-    );
-  });
-  const member = members.find(
-    (candidate) => candidate.id === queue.session_agent_id,
-  );
-
-  for (const item of activeQueueItems) {
-    if (String(item.message.status) !== 'processing') continue;
-    const sourceMessageId = item.message.chat_message_id;
-    const sourceMessage = next.find(
-      (message) =>
-        message.isUser &&
-        (message.id === sourceMessageId ||
-          userMessageClientId(message) === sourceMessageId),
-    );
-    const clientMessageId = sourceMessage
-      ? userMessageClientId(sourceMessage)
-      : undefined;
-    const hasMatchingPlaceholder = next.some(
-      (message) =>
-        message.isAgentRunning &&
-        message.sessionAgentId === queue.session_agent_id &&
-        (message.sourceMessageId === sourceMessageId ||
-          Boolean(
-            clientMessageId &&
-            message.clientMessageId === clientMessageId,
-          )),
-    );
-    if (hasMatchingPlaceholder) continue;
-
-    const sender = asAgentHandle(member?.name ?? 'agent');
-    next.push({
-      id: queueProcessingPlaceholderId(
-        sourceMessageId,
-        queue.session_agent_id,
-      ),
-      sessionId: queue.session_id,
-      avatar: member?.avatar ?? monogramFromName(sender),
-      sender,
-      model: member?.modelName,
-      time: 'just now',
-      createdAt:
-        item.message.processing_started_at ?? item.message.updated_at,
-      text: '',
-      isAgent: true,
-      isThinking: true,
-      isAgentRunning: true,
-      sourceMessageId,
-      clientMessageId,
-      sessionAgentId: queue.session_agent_id,
-    });
-  }
-
-  return orderMessagesForConversation(next);
-};
-
-const resolveProjectMainAgentMember = (
-  projectMembers: ProjectMemberWithRuntime[],
-): ProjectMemberWithRuntime | null =>
-  projectMembers.find(
-    (member) => member.member_type === 'agent' && member.role === 'lead',
-  ) ??
-  projectMembers.find((member) => member.member_type === 'agent') ??
-  null;
-
-const resolveProjectMainAgentId = (
-  projectMembers: ProjectMemberWithRuntime[],
-): string | null =>
-  resolveProjectMainAgentMember(projectMembers)?.agent_id ?? null;
-
-const resolveProjectMainAgentName = (
-  projectMembers: ProjectMemberWithRuntime[],
-  agents: BackendChatAgent[],
-): string | null => {
-  const mainMember = resolveProjectMainAgentMember(projectMembers);
-  if (!mainMember) return null;
-
-  const agent = mainMember.agent_id
-    ? agents.find((candidate) => candidate.id === mainMember.agent_id)
-    : undefined;
-  const displayName = mainMember.member_name?.trim() || agent?.name?.trim();
-  return displayName ? asAgentHandle(displayName) : null;
-};
-
-const withSessionId = (sessionId: string, message: Message): Message =>
-  message.sessionId === sessionId ? message : { ...message, sessionId };
-
-const withSessionIdsBySession = (
-  messagesBySession: Record<string, Message[]>,
-): Record<string, Message[]> =>
-  Object.fromEntries(
-    Object.entries(messagesBySession).map(([sessionId, messages]) => [
-      sessionId,
-      messages.map((message) => withSessionId(sessionId, message)),
-    ]),
-  );
-
-const filterMessagesForSession = (
-  sessionId: string,
-  messages: Message[],
-): Message[] => {
-  const scoped = messages.filter((message) => message.sessionId === sessionId);
-  const userIndexByClientId = new Map<string, number>();
-  const deduped: Message[] = [];
-
-  for (const message of scoped) {
-    if (message.isUser) {
-      const clientMessageId = userMessageClientId(message);
-      if (clientMessageId) {
-        const existingIndex = userIndexByClientId.get(clientMessageId);
-        if (existingIndex !== undefined) {
-          const existing = deduped[existingIndex];
-          deduped[existingIndex] =
-            isOptimisticUserMessage(existing) &&
-            !isOptimisticUserMessage(message)
-              ? message
-              : existing;
-          continue;
-        }
-        userIndexByClientId.set(clientMessageId, deduped.length);
-      }
-    }
-    deduped.push(message);
-  }
-
-  return orderMessagesForConversation(deduped);
-};
-
-export const mergePersistedWithRunningPlaceholders = (
-  persisted: Message[],
-  current: Message[],
-  activeSessionAgentIds?: Set<string>,
-  runningPlaceholders: Message[] = [],
-): Message[] => {
-  const correlated = correlateRunningPlaceholdersWithPending(
-    current,
-    runningPlaceholders,
-  );
-  const combinedCurrent = [
-    ...correlated.current,
-    ...correlated.runningPlaceholders,
-  ];
-  const persistedIds = new Set(persisted.map((message) => message.id));
-  const persistedClientMessageIds = new Set(
-    persisted
-      .map(userMessageClientId)
-      .filter((id): id is string => Boolean(id)),
-  );
-  const persistedRunIds = new Set(
-    persisted
-      .map((message) => message.runId)
-      .filter((runId): runId is string => Boolean(runId)),
-  );
-  const carriedMessagesByKey = new Map<string, Message>();
-  let hasRunIdPlaceholder = false;
-  for (const message of combinedCurrent) {
-    if (
-      message.isAgentRunning &&
-      message.sessionAgentId &&
-      activeSessionAgentIds &&
-      !activeSessionAgentIds.has(message.sessionAgentId) &&
-      !isOptimisticPendingAgentPlaceholder(message)
-    ) {
-      continue;
-    }
-
-    if (isOptimisticUserMessage(message)) {
-      const clientMessageId = userMessageClientId(message);
-      if (
-        !persistedIds.has(message.id) &&
-        clientMessageId &&
-        !persistedClientMessageIds.has(clientMessageId)
-      ) {
-        carriedMessagesByKey.set(`user:${clientMessageId}`, message);
-      }
-      continue;
-    }
-
-    if (!message.isAgentRunning || persistedIds.has(message.id)) continue;
-    if (message.runId && persistedRunIds.has(message.runId)) continue;
-    const pendingTargetKey =
-      message.sessionAgentId ?? normalizedAgentHandle(message.sender);
-    const key = `agent:${
-      message.runId ??
-      (message.clientMessageId
-        ? `${message.clientMessageId}:${pendingTargetKey}`
-        : message.id)
-    }`;
-    if (message.runId) hasRunIdPlaceholder = true;
-    const existing = carriedMessagesByKey.get(key);
-    carriedMessagesByKey.set(
-      key,
-      mergeCarriedRunPlaceholder(existing, message),
-    );
-  }
-
-  // If a real run placeholder exists, discard only hydrated pending placeholders
-  // (no runId). Keep optimistic pending placeholders because they can represent
-  // a newly queued message for the same agent while another run is active.
-  if (hasRunIdPlaceholder) {
-    for (const [key, message] of carriedMessagesByKey) {
-      if (
-        !message.runId &&
-        isPendingAgentPlaceholder(message) &&
-        !isOptimisticPendingAgentPlaceholder(message)
-      ) {
-        carriedMessagesByKey.delete(key);
-      }
-    }
-  }
-
-  const placeholders = [...carriedMessagesByKey.values()];
-  let merged = persisted;
-  for (const placeholder of placeholders) {
-    merged =
-      placeholder.sourceMessageId || placeholder.clientMessageId
-        ? [...merged, placeholder]
-        : insertMessageByCreatedAt(merged, placeholder);
-  }
-
-  return orderMessagesForConversation(merged);
-};
-
-const normalizeActiveRun = (run: ChatActiveRun): RuntimeActiveRun => run;
-
-const activeRunToMessage = (run: RuntimeActiveRun): Message => {
-  const displayName = run.display_name?.trim() || run.agent_name || 'agent';
-  const sender = displayName.startsWith('@') ? displayName : `@${displayName}`;
-  return {
-    id: `run-${run.run_id}`,
-    sessionId: run.session_id,
-    avatar: run.avatar || monogramFromName(displayName),
-    sender,
-    model: run.model ?? undefined,
-    time: 'just now',
-    createdAt: run.created_at,
-    text: '',
-    isAgent: true,
-    isThinking: true,
-    isAgentRunning: true,
-    runId: run.status === 'starting' ? undefined : run.run_id,
-    sessionAgentId: run.session_agent_id,
-    sourceMessageId: run.source_message_id ?? undefined,
-    clientMessageId: run.client_message_id ?? undefined,
-  };
-};
-
-const activeRunMessagesForSession = (
-  activeRunsByRunId: Record<string, RuntimeActiveRun>,
-  sessionId: string,
-): Message[] =>
-  Object.values(activeRunsByRunId)
-    .filter((run) => run.session_id === sessionId)
-    .map(activeRunToMessage);
-
-const mergeSessionMessagesWithActiveRuns = (
-  sessionMessages: Message[],
-  activeRunMessages: Message[],
-): Message[] =>
-  resolveMessageReferences(
-    mergePersistedWithRunningPlaceholders(
-      sessionMessages.filter((message) => !message.isAgentRunning),
-      sessionMessages,
-      undefined,
-      activeRunMessages,
-    ),
-  );
-
-export interface WorkspaceContextProps {
-  theme: Theme;
-  themePreference: ThemePreference;
-  setTheme: (t: ThemePreference) => void;
-  locale: Locale;
-  setLocale: (l: Locale) => void;
-  chatMessageFontSize: number;
-  setChatMessageFontSize: (size: number) => void;
-  members: Member[];
-  setMembers: (m: ListUpdater<Member>) => void;
-  sessions: Session[];
-  setSessions: (s: ListUpdater<Session>) => void;
-  projects: Project[];
-  projectsAsync: AsyncResourceState<Project[]>;
-  selectedProjectId: string;
-  setSelectedProjectId: (id: string) => void;
-  refreshProjects: () => Promise<void>;
-  createProject: (data: CreateProjectRequest) => Promise<Project>;
-  messages: Message[];
-  memberQueuesBySessionAgentId: MemberQueuesBySessionAgentId;
-  queuedUserMessagesById: Record<string, Message>;
-  workflowRuntimeLinesByExecution: Record<string, WorkflowRuntimeLine[]>;
-  activeSessionId: string;
-  setActiveSessionId: (id: string) => void;
-  chatInputMode: ChatInputMode;
-  setChatInputMode: (mode?: ChatInputMode) => void;
-  setSessionChatInputMode: (sessionId: string, mode: ChatInputMode) => void;
-  ensureWorkflowRouteToMainAgent: () => Promise<void>;
-  mainAgentName: string | null;
-  providers: Provider[];
-  setProviders: (p: ListUpdater<Provider>) => void;
-  strategies: Strategy[];
-  selectedStrategyId: string;
-  setSelectedStrategyId: (id: string) => void;
-  selectedOnboardType: 'saas' | 'cli' | 'game' | 'ai';
-  setSelectedOnboardType: (type: 'saas' | 'cli' | 'game' | 'ai') => void;
-  smartRouting: boolean;
-  setSmartRouting: (b: boolean) => void;
-  showCost: boolean;
-  setShowCost: (b: boolean) => void;
-  showExplanation: boolean;
-  setShowExplanation: (b: boolean) => void;
-  warnOverDollar: boolean;
-  setWarnOverDollar: (b: boolean) => void;
-  weeklyCost: number;
-  weeklySaved: number;
-  earlyBirdLeft: number;
-  setEarlyBirdLeft: (n: number | ((prev: number) => number)) => void;
-
-  // Modals state
-  isAddMemberModalOpen: boolean;
-  setIsAddMemberModalOpen: (b: boolean) => void;
-  isAddProviderModalOpen: boolean;
-  setIsAddProviderModalOpen: (b: boolean) => void;
-
-  // Active Simulation Utilities
-  sendMessage: (text: string, options?: SendMessageOptions) => void;
-  sendMessageToSession: (
-    sessionId: string,
-    text: string,
-    options?: SendMessageOptions,
-  ) => void;
-  addMemberToOrganization: (name: string, model: string) => void;
-  addProviderToKeychain: (name: string, key: string) => void;
-
-  // i18n hook helper
-  t: (key: string, replacements?: Record<string, string | number>) => string;
-
-  // Toast notifications
-  toast: WorkspaceToast | null;
-  showToast: (msg: string, tone?: ToastTone) => void;
-
-  // Settings active section
-  activeSettingsTab: string;
-  setActiveSettingsTab: (tab: string) => void;
-
-  // Async-status surface appended to the preserved legacy context shape.
-  sessionsAsync: AsyncResourceState<Session[]>;
-  refreshSessions: () => Promise<void>;
-  archivedSessionsAsync: AsyncResourceState<Session[]>;
-  refreshArchivedSessions: () => Promise<void>;
-  renameSession: (sessionId: string, title: string) => Promise<void>;
-  archiveSession: (sessionId: string) => Promise<void>;
-  pinSession: (sessionId: string, pinned: boolean) => Promise<void>;
-  deleteSession: (sessionId: string) => Promise<void>;
-  restoreSession: (sessionId: string) => Promise<void>;
-  messagesAsync: AsyncResourceState<Message[]>;
-  refreshMessages: () => Promise<void>;
-  /**
-   * Mark a stop request so the run does not keep session-level running
-   * indicators active while the visible placeholder stays in place until the
-   * persisted stopped message replaces it.
-   */
-  markSessionAgentStopped: (sessionAgentId: string) => void;
-  refreshMemberQueues: () => Promise<void>;
-  deleteQueuedMessage: (sessionId: string, queueId: string) => Promise<void>;
-  continueMemberQueue: (
-    sessionId: string,
-    sessionAgentId: string,
-  ) => Promise<void>;
-  membersAsync: AsyncResourceState<Member[]>;
-  refreshMembers: () => Promise<void>;
-  providersAsync: AsyncResourceState<Provider[]>;
-  refreshProviders: () => Promise<void>;
-  skills: BackendChatSkill[];
-  skillsAsync: AsyncResourceState<BackendChatSkill[]>;
-  refreshSkills: () => Promise<void>;
-  config: Config | null;
-  configAsync: AsyncResourceState<Config | null>;
-  refreshConfig: () => Promise<void>;
-  saveConfigPatch: (patch: Partial<Config>) => Promise<Config>;
-  environment: Environment | null;
-  inboxSummaryAsync: AsyncResourceState<InboxSummary>;
-  inboxItemsAsync: AsyncResourceState<InboxItem[]>;
-  refreshInbox: () => Promise<void>;
-  markInboxItemRead: (itemId: string) => Promise<void>;
-  markInboxItemsRead: (itemIds: string[]) => Promise<void>;
-  markAllInboxRead: () => Promise<void>;
-  archiveInboxItem: (itemId: string) => Promise<void>;
-  workflowCard: WorkflowCardProjection | null;
-  workflowCardAsync: AsyncResourceState<WorkflowCardProjection | null>;
-  refreshWorkflowCard: (messageId: string) => Promise<void>;
-  refreshSessionWorkflowStatus: (sessionId: string) => Promise<void>;
-  workspaceChanges: WorkspaceChangesResponse | null;
-  workspaceChangesAsync: AsyncResourceState<WorkspaceChangesResponse | null>;
-  refreshWorkspaceChanges: (
-    sessionId: string,
-    path: string,
-    includeDiff?: boolean,
-  ) => Promise<void>;
-  resetWorkspaceChanges: () => void;
-  /** Re-runs every auto-loaded resource. Useful as a global retry. */
-  refreshAll: () => Promise<void>;
-}
+import { useWorkspaceState } from './workspace/useWorkspaceState';
+import { useWorkspaceChatRuntime } from './workspace/useWorkspaceChatRuntime';
+import type { WorkspaceContextProps } from './workspace/workspaceContextContract';
+import type { ChatStreamEvent } from './workspace/workspaceChatStreamTypes';
+export type { WorkspaceContextProps } from './workspace/workspaceContextContract';
+import {
+  DEFAULT_CHAT_INPUT_MODE,
+  chatSessionUpdatePayload,
+  resolveChatInputMode,
+  type SendMessageOptions,
+} from './workspace/workspaceContextTypes';
+export * from './workspace/workspaceContextTypes';
+import {
+  EMPTY_INBOX_SUMMARY,
+  INBOX_LIGHT_REFRESH_DELAY_MS,
+  INBOX_LIST_LIMIT,
+  INBOX_REFRESH_INTERVAL_MS,
+  SIDEBAR_RUNNING_INDICATOR_POLL_MS,
+  STARTING_AGENT_RECONCILE_DELAY_MS,
+  activeRunMessagesForSession,
+  activeRunToMessage,
+  chatMessageFontSizeFromConfig,
+  countUnreadInboxItems,
+  createClientMessageId,
+  decrementInboxSummaryEntries,
+  extractAgentMentions,
+  filterInboxItemsForEnabledSources,
+  filterInboxSummaryForEnabledSources,
+  filterMessagesForSession,
+  filterQueuedUserMessagesFromSnapshot,
+  hasRunningSessionAgent,
+  inboxCountValue,
+  inboxNotificationSettingsSignature,
+  isAutoReadableInboxItem,
+  isOptimisticPendingAgentPlaceholder,
+  isPendingAgentPlaceholder,
+  loadSessionRunningIndicators,
+  localeFromConfig,
+  makePendingAgentPlaceholders,
+  mergePersistedWithRunningPlaceholders,
+  mergeSessionMessagesWithActiveRuns,
+  normalizeActiveRun,
+  playInboxNotificationSound,
+  queuedUserMessagesByIdFromSnapshot,
+  resolveProjectMainAgentId,
+  resolveProjectMainAgentName,
+  showInboxSystemNotification,
+  themePreferenceFromConfig,
+  userMessageClientId,
+  withSessionIdsBySession,
+} from './workspace/workspaceContextUtils';
+
+export type { ChatStreamEvent } from './workspace/workspaceChatStreamTypes';
+export * from './workspace/workspaceContextUtils';
 
 export const WorkspaceContext = createContext<WorkspaceContextProps | undefined>(
   undefined,
@@ -1811,950 +111,161 @@ export const WorkspaceContext = createContext<WorkspaceContextProps | undefined>
 export const WorkspaceProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  const runActivityStore = useRunActivityStore();
-  const [themePreference, setThemePreferenceState] =
-    useState<ThemePreference>('system');
-  const [systemTheme, setSystemTheme] = useState<Theme>(resolveSystemTheme);
-  const theme: Theme =
-    themePreference === 'system' ? systemTheme : themePreference;
-
-  const [locale, setLocaleState] = useState<Locale>(resolveBrowserLocale);
-  const [chatMessageFontSize, setChatMessageFontSizeState] =
-    useState<number>(CHAT_MESSAGE_FONT_SIZE_DEFAULT);
-  const [activeSessionId, setActiveSessionId] = useState<string>(() => {
-    try {
-      return localStorage.getItem(ACTIVE_SESSION_ID_STORAGE_KEY) ?? '';
-    } catch {
-      return '';
-    }
-  });
-  const mockBootstrapRef = useRef<WorkspaceBootstrapMock | null>(null);
-  const toastDurationMsRef = useRef(3000);
-
-  // Async-backed primary resources. Each is seeded with the existing mock so
-  // the UI renders before the first API response arrives (or if the backend
-  // is unreachable / has a contract gap).
-  const [sessionsAsync, setSessionsAsync] = useState<
-    AsyncResourceState<Session[]>
-  >(() => initialAsync([]));
-  const [archivedSessionsAsync, setArchivedSessionsAsync] = useState<
-    AsyncResourceState<Session[]>
-  >(() => initialAsync([]));
-  const [projectsAsync, setProjectsAsync] = useState<
-    AsyncResourceState<Project[]>
-  >(() => initialAsync([]));
-  const [selectedProjectId, setSelectedProjectIdState] = useState<string>(() => {
-    try {
-      return localStorage.getItem(SELECTED_PROJECT_ID_STORAGE_KEY) ?? '';
-    } catch {
-      return '';
-    }
-  });
-  const [allMessages, setAllMessages] = useState<Record<string, Message[]>>({});
-  const allMessagesRef = useRef<Record<string, Message[]>>({});
-  const [memberQueuesBySessionAgentId, setMemberQueuesBySessionAgentId] =
-    useState<MemberQueuesBySessionAgentId>({});
-  const [activeRunsByRunId, setActiveRunsByRunId] = useState<
-    Record<string, RuntimeActiveRun>
-  >({});
-  const [workflowRuntimeLinesByExecution, setWorkflowRuntimeLinesByExecution] =
-    useState<Record<string, WorkflowRuntimeLine[]>>({});
-  const [messagesAsync, setMessagesAsync] = useState<
-    AsyncResourceState<Message[]>
-  >(() => initialAsync([]));
-  const [membersAsync, setMembersAsync] = useState<
-    AsyncResourceState<Member[]>
-  >(() => initialAsync([]));
-  const [mainAgentName, setMainAgentName] = useState<string | null>(null);
-  const [providersAsync, setProvidersAsync] = useState<
-    AsyncResourceState<Provider[]>
-  >(() => initialAsync([]));
-  const [skillsAsync, setSkillsAsync] = useState<
-    AsyncResourceState<BackendChatSkill[]>
-  >(() => initialAsync([]));
-  const [configAsync, setConfigAsync] = useState<
-    AsyncResourceState<Config | null>
-  >(() => initialAsync(null));
-  const [environment, setEnvironment] = useState<Environment | null>(null);
-  const latestConfigRef = useRef<Config | null>(null);
-  const configPatchQueueRef = useRef<ReturnType<
-    typeof createConfigPatchQueue<Config>
-  > | null>(null);
-  const publishVisibleConfig = useCallback((visible: Config) => {
-    latestConfigRef.current = visible;
-    setConfigAsync(succeed(visible));
-  }, []);
-  const ensureConfigPatchQueue = useCallback(
-    (initial: Config) => {
-      if (!configPatchQueueRef.current) {
-        configPatchQueueRef.current = createConfigPatchQueue<Config>(
-          initial,
-          systemApi.saveConfig,
-          publishVisibleConfig,
-        );
-      } else {
-        configPatchQueueRef.current.replaceAcknowledged(initial);
-      }
-      return configPatchQueueRef.current;
-    },
-    [publishVisibleConfig],
-  );
-  const saveConfigPatch = useCallback((patch: Partial<Config>) => {
-    const queue = configPatchQueueRef.current;
-    if (!queue) return Promise.reject(new Error('Config is not loaded'));
-    return queue.enqueue(patch, { optimistic: false });
-  }, []);
-  const [inboxSummaryAsync, setInboxSummaryAsync] = useState<
-    AsyncResourceState<InboxSummary>
-  >(() => initialAsync(EMPTY_INBOX_SUMMARY));
-  const [inboxItemsAsync, setInboxItemsAsync] = useState<
-    AsyncResourceState<InboxItem[]>
-  >(() => initialAsync([]));
-  const [workflowCardAsync, setWorkflowCardAsync] = useState<
-    AsyncResourceState<WorkflowCardProjection | null>
-  >(() => initialAsync(null));
-  const [workspaceChangesAsync, setWorkspaceChangesAsync] = useState<
-    AsyncResourceState<WorkspaceChangesResponse | null>
-  >(() => initialAsync(null));
-  const messagesRequestIdRef = useRef(0);
-  const queueRequestIdRef = useRef(0);
-  const inboxRequestIdRef = useRef(0);
-  const inboxLightRefreshTimerRef = useRef<ReturnType<typeof setTimeout> | null>(
-    null,
-  );
-  const inboxSoundProjectIdRef = useRef<string | null>(null);
-  const inboxSoundSettingsSignatureRef = useRef<string | null>(null);
-  const inboxSoundPrimedRef = useRef(false);
-  const inboxUnreadSoundIdsRef = useRef<Set<string>>(new Set());
-  const inboxAutoReadProjectIdRef = useRef<string | null>(null);
-  const inboxInitialUnreadItemIdsRef = useRef<Set<string>>(new Set());
-  const autoMarkedInboxItemIdsRef = useRef<Set<string>>(new Set());
-  const workspaceChangesRequestIdRef = useRef(0);
-  const initialRefreshStartedRef = useRef(false);
-  const initialRefreshCompletedRef = useRef(false);
-  const sessionRunningIndicatorRequestsRef = useRef<Map<string, Promise<void>>>(
-    new Map(),
-  );
-  const sessionWorkflowStatusRequestsRef = useRef<
-    Map<string, Promise<WorkflowSessionStatusResponse | null>>
-  >(new Map());
-  const [chatInputModeBySessionId, setChatInputModeBySessionId] = useState<
-    Record<string, ChatInputMode>
-  >({});
-
-  const [strategies, setStrategies] = useState<Strategy[]>([]);
-  const [mockAgentRepliesByMention, setMockAgentRepliesByMention] = useState<
-    Record<string, string[]>
-  >({ default: ['Working on it.'] });
-  const [selectedStrategyId, setSelectedStrategyId] = useState<string>('');
-  const [selectedOnboardType, setSelectedOnboardType] = useState<
-    'saas' | 'cli' | 'game' | 'ai'
-  >('saas');
-
-  // Global Settings Switches
-  const [smartRouting, setSmartRouting] = useState<boolean>(true);
-  const [showCost, setShowCost] = useState<boolean>(true);
-  const [showExplanation, setShowExplanation] = useState<boolean>(true);
-  const [warnOverDollar, setWarnOverDollar] = useState<boolean>(false);
-
-  // Stats (LOCAL / MOCK-FALLBACK per backend_contract_audit 搂5.1)
-  const [weeklyCost, setWeeklyCost] = useState<number>(0);
-  const [weeklySaved, setWeeklySaved] = useState<number>(0);
-  const [earlyBirdLeft, setEarlyBirdLeft] = useState<number>(0);
-
-  // Settings view controller
-  const [activeSettingsTab, setActiveSettingsTab] =
-    useState<string>('providers');
-
-  // Modal Switches
-  const [isAddMemberModalOpen, setIsAddMemberModalOpen] =
-    useState<boolean>(false);
-  const [isAddProviderModalOpen, setIsAddProviderModalOpen] =
-    useState<boolean>(false);
-
-  // Toast
-  const [toast, setToast] = useState<WorkspaceToast | null>(null);
-
-  // Cache the latest activeSessionId so async callbacks see the live value.
-  const activeSessionIdRef = useRef(activeSessionId);
-  const selectedProjectIdRef = useRef(selectedProjectId);
-  // Cache the active session's workspace path so the WebSocket
-  // `file_change_refresh` handler can refresh workspace changes without a stale
-  // closure (the socket effect does not re-subscribe on every sessions update).
-  const activeWorkspacePathRef = useRef<string | null>(null);
-  const sessionLeadAgentIdBySessionIdRef = useRef<Record<
-    string,
-    string | null
-  >>({});
-  const workflowRouteAgentIdRef = useRef<string | null>(null);
-  const agentNamesByIdRef = useRef<Record<string, string>>({});
-  const agentModelsByIdRef = useRef<Record<string, string | null>>({});
-  const notifiedTokenUsageSignaturesRef = useRef<Record<string, string>>({});
-  // Session agents the user has just requested to stop. While an agent is in
-  // this set, keep any existing visible placeholder until the persisted stop
-  // notice replaces it, but do not re-hydrate a separate running placeholder.
-  // Cleared when a new run starts or after the terminal stop notice replaces
-  // the placeholder.
-  const optimisticallyStoppedSessionAgentIdsRef = useRef<Set<string>>(
-    new Set(),
-  );
-  const runningAgentSessionIdsRef = useRef<Set<string>>(
-    readSessionIdSet(RUNNING_AGENT_SESSION_IDS_STORAGE_KEY),
-  );
-  const unreadAgentCompletionSessionIdsRef = useRef<Set<string>>(
-    readSessionIdSet(UNREAD_AGENT_COMPLETION_SESSION_IDS_STORAGE_KEY),
-  );
-  const acknowledgedWorkflowInputIdsRef = useRef<Set<string>>(
-    readSessionIdSet(ACKED_WORKFLOW_INPUT_IDS_STORAGE_KEY),
-  );
-  const acknowledgedWorkflowErrorSessionIdsRef = useRef<Set<string>>(
-    readSessionIdSet(ACKED_WORKFLOW_ERROR_SESSION_IDS_STORAGE_KEY),
-  );
-  useEffect(() => {
-    allMessagesRef.current = allMessages;
-  }, [allMessages]);
-
-  const persistAgentSessionActivityStorage = useCallback(() => {
-    writeSessionIdSet(
-      RUNNING_AGENT_SESSION_IDS_STORAGE_KEY,
-      runningAgentSessionIdsRef.current,
-    );
-    writeSessionIdSet(
-      UNREAD_AGENT_COMPLETION_SESSION_IDS_STORAGE_KEY,
-      unreadAgentCompletionSessionIdsRef.current,
-    );
-  }, []);
-  const persistWorkflowInputAcknowledgementStorage = useCallback(() => {
-    writeSessionIdSet(
-      ACKED_WORKFLOW_INPUT_IDS_STORAGE_KEY,
-      acknowledgedWorkflowInputIdsRef.current,
-    );
-  }, []);
-  const persistWorkflowErrorAcknowledgementStorage = useCallback(() => {
-    writeSessionIdSet(
-      ACKED_WORKFLOW_ERROR_SESSION_IDS_STORAGE_KEY,
-      acknowledgedWorkflowErrorSessionIdsRef.current,
-    );
-  }, []);
-
-  const syncSessionAgentActivityIndicator = useCallback(
-    (sessionId: string, hasRunningAgent: boolean): boolean => {
-      if (!sessionId) return false;
-
-      let changed = false;
-      if (hasRunningAgent) {
-        if (!runningAgentSessionIdsRef.current.has(sessionId)) {
-          runningAgentSessionIdsRef.current.add(sessionId);
-          changed = true;
-        }
-        if (unreadAgentCompletionSessionIdsRef.current.delete(sessionId)) {
-          changed = true;
-        }
-      } else {
-        const wasRunning = runningAgentSessionIdsRef.current.delete(sessionId);
-        if (wasRunning) {
-          changed = true;
-        }
-        if (activeSessionIdRef.current === sessionId) {
-          if (unreadAgentCompletionSessionIdsRef.current.delete(sessionId)) {
-            changed = true;
-          }
-        } else if (wasRunning) {
-          unreadAgentCompletionSessionIdsRef.current.add(sessionId);
-          changed = true;
-        }
-      }
-
-      if (changed) {
-        persistAgentSessionActivityStorage();
-      }
-      return unreadAgentCompletionSessionIdsRef.current.has(sessionId);
-    },
-    [persistAgentSessionActivityStorage],
-  );
-
-  const acknowledgeWorkflowInput = useCallback(
-    (inputId: string | null | undefined) => {
-      if (!inputId || acknowledgedWorkflowInputIdsRef.current.has(inputId)) {
-        return;
-      }
-      acknowledgedWorkflowInputIdsRef.current.add(inputId);
-      persistWorkflowInputAcknowledgementStorage();
-    },
-    [persistWorkflowInputAcknowledgementStorage],
-  );
-
-  const syncSessionWorkflowInputIndicator = useCallback(
-    (
-      sessionId: string,
-      pendingWorkflowInputId: string | null | undefined,
-    ): boolean => {
-      if (!sessionId || !pendingWorkflowInputId) return false;
-      if (activeSessionIdRef.current === sessionId) {
-        acknowledgeWorkflowInput(pendingWorkflowInputId);
-        return false;
-      }
-      return !acknowledgedWorkflowInputIdsRef.current.has(
-        pendingWorkflowInputId,
-      );
-    },
-    [acknowledgeWorkflowInput],
-  );
-
-  const acknowledgeWorkflowError = useCallback(
-    (sessionId: string | null | undefined) => {
-      if (!sessionId) return;
-      if (acknowledgedWorkflowErrorSessionIdsRef.current.has(sessionId)) {
-        return;
-      }
-      acknowledgedWorkflowErrorSessionIdsRef.current.add(sessionId);
-      persistWorkflowErrorAcknowledgementStorage();
-    },
-    [persistWorkflowErrorAcknowledgementStorage],
-  );
-
-  const syncSessionWorkflowErrorIndicator = useCallback(
-    (sessionId: string, workflowSidebarState: WorkflowSidebarState): boolean => {
-      if (!sessionId) return false;
-      if (workflowSidebarState !== 'failed') {
-        if (acknowledgedWorkflowErrorSessionIdsRef.current.delete(sessionId)) {
-          persistWorkflowErrorAcknowledgementStorage();
-        }
-        return false;
-      }
-      if (activeSessionIdRef.current === sessionId) {
-        acknowledgeWorkflowError(sessionId);
-        return false;
-      }
-      return !acknowledgedWorkflowErrorSessionIdsRef.current.has(sessionId);
-    },
-    [acknowledgeWorkflowError, persistWorkflowErrorAcknowledgementStorage],
-  );
-
-  const clearUnreadAgentCompletion = useCallback(
-    (sessionId: string) => {
-      if (!sessionId) return;
-      if (!unreadAgentCompletionSessionIdsRef.current.delete(sessionId)) {
-        return;
-      }
-
-      persistAgentSessionActivityStorage();
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (
-            session.id !== sessionId ||
-            !session.hasUnreadAgentCompletion
-          ) {
-            return session;
-          }
-          changed = true;
-          return { ...session, hasUnreadAgentCompletion: false };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [persistAgentSessionActivityStorage],
-  );
-
-  const clearPendingWorkflowInput = useCallback(
-    (sessionId: string) => {
-      if (!sessionId) return;
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (session.id !== sessionId) return session;
-          acknowledgeWorkflowInput(session.pendingWorkflowInputId);
-          if (!session.hasPendingWorkflowInput) return session;
-          changed = true;
-          return { ...session, hasPendingWorkflowInput: false };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [acknowledgeWorkflowInput],
-  );
-
-  const clearWorkflowErrorAttention = useCallback(
-    (sessionId: string) => {
-      if (!sessionId) return;
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (session.id !== sessionId) {
-            return session;
-          }
-          if (session.workflowSidebarState === 'failed') {
-            acknowledgeWorkflowError(sessionId);
-          }
-          if (!session.hasWorkflowError) return session;
-          changed = true;
-          return { ...session, hasWorkflowError: false };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [acknowledgeWorkflowError],
-  );
-
-  useEffect(() => {
-    activeSessionIdRef.current = activeSessionId;
-    try {
-      if (activeSessionId) {
-        localStorage.setItem(ACTIVE_SESSION_ID_STORAGE_KEY, activeSessionId);
-      } else {
-        localStorage.removeItem(ACTIVE_SESSION_ID_STORAGE_KEY);
-      }
-    } catch {}
-    clearUnreadAgentCompletion(activeSessionId);
-    clearPendingWorkflowInput(activeSessionId);
-    clearWorkflowErrorAttention(activeSessionId);
-  }, [
+  const workspaceState = useWorkspaceState();
+  const {
+    themePreference,
+    setThemePreferenceState,
+    systemTheme,
+    setSystemTheme,
+    locale,
+    setLocaleState,
+    chatMessageFontSize,
+    setChatMessageFontSizeState,
     activeSessionId,
-    clearPendingWorkflowInput,
+    setActiveSessionId,
+    sessionsAsync,
+    setSessionsAsync,
+    archivedSessionsAsync,
+    setArchivedSessionsAsync,
+    projectsAsync,
+    setProjectsAsync,
+    selectedProjectId,
+    setSelectedProjectIdState,
+    allMessages,
+    setAllMessages,
+    memberQueuesBySessionAgentId,
+    setMemberQueuesBySessionAgentId,
+    activeRunsByRunId,
+    setActiveRunsByRunId,
+    workflowRuntimeLinesByExecution,
+    setWorkflowRuntimeLinesByExecution,
+    messagesAsync,
+    setMessagesAsync,
+    membersAsync,
+    setMembersAsync,
+    mainAgentName,
+    setMainAgentName,
+    providersAsync,
+    setProvidersAsync,
+    skillsAsync,
+    setSkillsAsync,
+    configAsync,
+    setConfigAsync,
+    environment,
+    setEnvironment,
+    inboxSummaryAsync,
+    setInboxSummaryAsync,
+    inboxItemsAsync,
+    setInboxItemsAsync,
+    workflowCardAsync,
+    setWorkflowCardAsync,
+    workspaceChangesAsync,
+    setWorkspaceChangesAsync,
+    chatInputModeBySessionId,
+    setChatInputModeBySessionId,
+    strategies,
+    setStrategies,
+    mockAgentRepliesByMention,
+    setMockAgentRepliesByMention,
+    selectedStrategyId,
+    setSelectedStrategyId,
+    selectedOnboardType,
+    setSelectedOnboardType,
+    smartRouting,
+    setSmartRouting,
+    showCost,
+    setShowCost,
+    showExplanation,
+    setShowExplanation,
+    warnOverDollar,
+    setWarnOverDollar,
+    weeklyCost,
+    setWeeklyCost,
+    weeklySaved,
+    setWeeklySaved,
+    earlyBirdLeft,
+    setEarlyBirdLeft,
+    activeSettingsTab,
+    setActiveSettingsTab,
+    isAddMemberModalOpen,
+    setIsAddMemberModalOpen,
+    isAddProviderModalOpen,
+    setIsAddProviderModalOpen,
+    toast,
+    setToast,
+    runActivityStore,
+    theme,
+    mockBootstrapRef,
+    toastDurationMsRef,
+    allMessagesRef,
+    latestConfigRef,
+    configPatchQueueRef,
+    publishVisibleConfig,
+    ensureConfigPatchQueue,
+    saveConfigPatch,
+    messagesRequestIdRef,
+    queueRequestIdRef,
+    inboxRequestIdRef,
+    inboxLightRefreshTimerRef,
+    inboxSoundProjectIdRef,
+    inboxSoundSettingsSignatureRef,
+    inboxSoundPrimedRef,
+    inboxUnreadSoundIdsRef,
+    inboxAutoReadProjectIdRef,
+    inboxInitialUnreadItemIdsRef,
+    autoMarkedInboxItemIdsRef,
+    workspaceChangesRequestIdRef,
+    initialRefreshStartedRef,
+    initialRefreshCompletedRef,
+    sessionRunningIndicatorRequestsRef,
+    sessionWorkflowStatusRequestsRef,
+    activeSessionIdRef,
+    selectedProjectIdRef,
+    activeWorkspacePathRef,
+    sessionLeadAgentIdBySessionIdRef,
+    workflowRouteAgentIdRef,
+    agentNamesByIdRef,
+    agentModelsByIdRef,
+    notifiedTokenUsageSignaturesRef,
+    optimisticallyStoppedSessionAgentIdsRef,
+    runningAgentSessionIdsRef,
+    unreadAgentCompletionSessionIdsRef,
+    acknowledgedWorkflowInputIdsRef,
+    acknowledgedWorkflowErrorSessionIdsRef,
+    persistAgentSessionActivityStorage,
+    persistWorkflowInputAcknowledgementStorage,
+    persistWorkflowErrorAcknowledgementStorage,
+    syncSessionAgentActivityIndicator,
+    acknowledgeWorkflowInput,
+    syncSessionWorkflowInputIndicator,
+    acknowledgeWorkflowError,
+    syncSessionWorkflowErrorIndicator,
     clearUnreadAgentCompletion,
+    clearPendingWorkflowInput,
     clearWorkflowErrorAttention,
-  ]);
-
-  useEffect(() => {
-    messagesRequestIdRef.current += 1;
-    const sessionQueues = activeSessionId
-      ? Object.values(memberQueuesBySessionAgentId).filter(
-          (queue) => queue.session_id === activeSessionId,
-        )
-      : [];
-    const sessionMessages = activeSessionId
-      ? filterMessagesForSession(
-          activeSessionId,
-          allMessagesRef.current[activeSessionId] ?? [],
-        )
-      : [];
-    const sessionActiveRunMessages = activeSessionId
-      ? activeRunMessagesForSession(activeRunsByRunId, activeSessionId)
-      : [];
-    const sessionSnapshot = activeSessionId
-      ? mergeSessionMessagesWithActiveRuns(
-          sessionMessages,
-          sessionActiveRunMessages,
-        )
-      : [];
-    setMessagesAsync(
-      succeed(
-        activeSessionId
-          ? filterQueuedUserMessagesFromSnapshot(
-              sessionSnapshot,
-              sessionQueues,
-              activeSessionId,
-            )
-          : [],
-      ),
-    );
-  }, [activeRunsByRunId, activeSessionId, memberQueuesBySessionAgentId]);
-
-  // Keep the cached workspace path in sync with the active session so the
-  // WebSocket `file_change_refresh` handler always refreshes the right path.
-  // Mirrors FreeChatWorkspace's `reloadRelatedFiles` (currentProject workspace).
-  useEffect(() => {
-    activeWorkspacePathRef.current = selectedProjectId
-      ? projectsAsync.data?.find(
-          (project) => project.id === selectedProjectId,
-        )?.default_workspace_path ?? null
-      : null;
-  }, [selectedProjectId, projectsAsync]);
-  useEffect(() => {
-    setWorkflowRuntimeLinesByExecution({});
-  }, [activeSessionId]);
-  useEffect(() => {
-    selectedProjectIdRef.current = selectedProjectId;
-  }, [selectedProjectId]);
-
-  const chatInputMode =
-    activeSessionId !== ''
-      ? (chatInputModeBySessionId[activeSessionId] ??
-        DEFAULT_CHAT_INPUT_MODE)
-      : DEFAULT_CHAT_INPUT_MODE;
-
-  const showToast = (msg: string, tone: ToastTone = 'info') => {
-    setToast({ message: msg, tone });
-    setTimeout(() => {
-      setToast(null);
-    }, toastDurationMsRef.current);
-  };
-
-  const persistUiPreference = (patch: Partial<Config>) => {
-    const queue = configPatchQueueRef.current;
-    if (!queue) {
-      showToast('Settings are still loading. Please try again.', 'error');
-      return;
-    }
-    void queue.enqueue(patch, { optimistic: true }).catch((error) => {
-      console.error('Failed to save UI preferences', error);
-      showToast('Failed to save settings to config.json.', 'error');
-    });
-  };
-
-  const setTheme = (t: ThemePreference) => {
-    setThemePreferenceState(t);
-    persistUiPreference({ theme: themePreferenceToConfig(t) });
-  };
-
-  const setLocale = (l: Locale) => {
-    setLocaleState(l);
-    persistUiPreference({ language: localeToConfig(l) });
-  };
-
-  const setChatMessageFontSize = (size: number) => {
-    const normalized = normalizeChatMessageFontSize(size);
-    setChatMessageFontSizeState(normalized);
-    persistUiPreference({
-      chat_bubble_font_size: chatMessageFontSizeToConfig(normalized),
-    });
-  };
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    const mediaQuery = window.matchMedia?.('(prefers-color-scheme: light)');
-    if (!mediaQuery) return;
-
-    const updateSystemTheme = () => {
-      setSystemTheme(mediaQuery.matches ? 'light' : 'dark');
-    };
-
-    updateSystemTheme();
-    mediaQuery.addEventListener?.('change', updateSystemTheme);
-    return () => {
-      mediaQuery.removeEventListener?.('change', updateSystemTheme);
-    };
-  }, []);
-
-  useEffect(() => {
-    document.body.setAttribute('data-mode', theme);
-  }, [theme]);
-
-  const makeListSetter =
-    <T,>(
-      setAsync: React.Dispatch<React.SetStateAction<AsyncResourceState<T[]>>>,
-    ) =>
-    (next: ListUpdater<T>) => {
-      setAsync((prev) => {
-        const newData =
-          typeof next === 'function'
-            ? (next as (p: T[]) => T[])(prev.data)
-            : next;
-        return { ...prev, data: newData, empty: newData.length === 0 };
-      });
-    };
-
-  const setSessions = useCallback(
-    makeListSetter<Session>(setSessionsAsync),
-    [],
-  );
-  const setMembers = useCallback(makeListSetter<Member>(setMembersAsync), []);
-  const setProviders = useCallback(
-    makeListSetter<Provider>(setProvidersAsync),
-    [],
-  );
-  const setSessionRunningIndicator = useCallback(
-    (sessionId: string, hasRunningAgent: boolean) => {
-      if (!sessionId) return;
-      const hasUnreadAgentCompletion = syncSessionAgentActivityIndicator(
-        sessionId,
-        hasRunningAgent,
-      );
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (
-            session.id !== sessionId ||
-            (session.hasRunningAgent === hasRunningAgent &&
-              session.hasUnreadAgentCompletion === hasUnreadAgentCompletion)
-          ) {
-            return session;
-          }
-          changed = true;
-          return {
-            ...session,
-            hasRunningAgent,
-            hasUnreadAgentCompletion,
-          };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [syncSessionAgentActivityIndicator],
-  );
-  const syncProcessingQueuePlaceholders = useCallback(
-    (sessionId: string, queues: ReadonlyArray<MemberQueueSnapshot>) => {
-      setAllMessages((prev) => {
-        const current = filterMessagesForSession(
-          sessionId,
-          prev[sessionId] ?? [],
-        );
-        let next = current;
-        for (const queue of queues) {
-          if (queue.session_id !== sessionId) continue;
-          next = reconcileProcessingQueuePlaceholders(
-            next,
-            queue,
-            membersAsync.data,
-          );
-        }
-        const unchanged =
-          next.length === current.length &&
-          next.every((message, index) => message === current[index]);
-        return unchanged ? prev : { ...prev, [sessionId]: next };
-      });
-    },
-    [membersAsync.data],
-  );
-  const applyChatRuntimeSnapshot = useCallback(
-    (snapshot: ChatSessionRuntimeSnapshot) => {
-      const sid = snapshot.session_id;
-      setMemberQueuesBySessionAgentId((prev) => {
-        const next = { ...prev };
-        for (const [sessionAgentId, queue] of Object.entries(next)) {
-          if (queue.session_id === sid) {
-            delete next[sessionAgentId];
-          }
-        }
-        for (const queue of snapshot.queues) {
-          next[queue.session_agent_id] = queue;
-        }
-        return next;
-      });
-      syncProcessingQueuePlaceholders(sid, snapshot.queues);
-      setActiveRunsByRunId((prev) => {
-        const next = { ...prev };
-        const existingSessionRuns = Object.values(prev).filter(
-          (run) => run.session_id === sid,
-        );
-        for (const [runId, run] of Object.entries(next)) {
-          if (run.session_id === sid) {
-            delete next[runId];
-          }
-        }
-        for (const run of snapshot.active_runs) {
-          const existingRealRun = existingSessionRuns.find(
-            (candidate) =>
-              candidate.session_agent_id === run.session_agent_id &&
-              candidate.status !== 'starting' &&
-              (candidate.client_message_id === run.client_message_id ||
-                !run.client_message_id),
-          );
-          if (run.status === 'starting' && existingRealRun) {
-            next[existingRealRun.run_id] = existingRealRun;
-            continue;
-          }
-          next[run.run_id] = normalizeActiveRun(run);
-        }
-        return next;
-      });
-      runActivityStore.syncRuns(
-        snapshot.active_runs
-          .filter((run) => run.status !== 'starting')
-          .map((run) => run.run_id),
-      );
-      setSessionRunningIndicator(sid, snapshot.active_runs.length > 0);
-      if (snapshot.messages) {
-        const mapped = mapMessages(snapshot.messages as unknown as BackendChatMessage[], {
-          agentNamesById: agentNamesByIdRef.current,
-          agentModelsById: agentModelsByIdRef.current,
-        });
-        const activeRuns = snapshot.active_runs.map(normalizeActiveRun);
-        const activeSessionAgentIds = new Set(
-          activeRuns.map((run) => run.session_agent_id),
-        );
-        const runningPlaceholders = activeRuns.map(activeRunToMessage);
-        setAllMessages((prev) => {
-          const current = filterMessagesForSession(sid, prev[sid] ?? []);
-          return {
-            ...prev,
-            [sid]: resolveMessageReferences(
-              mergePersistedWithRunningPlaceholders(
-                mapped,
-                current,
-                activeSessionAgentIds,
-                runningPlaceholders,
-              ),
-            ),
-          };
-        });
-      }
-    },
-    [
-      runActivityStore,
-      setSessionRunningIndicator,
-      syncProcessingQueuePlaceholders,
-    ],
-  );
-  const reconcileStartingPlaceholders = useCallback(
-    async (sessionId: string, clientMessageId: string): Promise<void> => {
-      let snapshot: ChatSessionRuntimeSnapshot;
-      try {
-        snapshot = await chatRuntimeApi.getSnapshot(sessionId);
-      } catch {
-        return;
-      }
-      applyChatRuntimeSnapshot(snapshot);
-
-      setAllMessages((prev) => {
-        const current = filterMessagesForSession(
-          sessionId,
-          prev[sessionId] ?? [],
-        );
-        let changed = false;
-        const next = current.filter((message) => {
-          if (
-            !isPendingAgentPlaceholder(message) ||
-            message.clientMessageId !== clientMessageId
-          ) {
-            return true;
-          }
-          const hasActiveRun = snapshot.active_runs.some(
-            (run) =>
-              message.sessionAgentId
-                ? run.session_agent_id === message.sessionAgentId
-                : normalizedAgentHandle(run.display_name || run.agent_name) ===
-                  normalizedAgentHandle(message.sender),
-          );
-          const isQueued = isStartingPlaceholderRepresentedInQueue(
-            message,
-            snapshot.queues,
-            sessionId,
-          );
-          const keep = hasActiveRun || isQueued;
-          if (!keep) changed = true;
-          return keep;
-        });
-        return changed ? { ...prev, [sessionId]: next } : prev;
-      });
-    },
-    [applyChatRuntimeSnapshot],
-  );
-  const setSessionWorkflowRunningIndicator = useCallback(
-    (sessionId: string, hasRunningWorkflow: boolean) => {
-      if (!sessionId) return;
-      const workflowSidebarState: WorkflowSidebarState = hasRunningWorkflow
-        ? 'running'
-        : 'idle';
-      const hasWorkflowError = syncSessionWorkflowErrorIndicator(
-        sessionId,
-        workflowSidebarState,
-      );
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (
-            session.id !== sessionId ||
-            (session.hasRunningWorkflow === hasRunningWorkflow &&
-              session.workflowSidebarState === workflowSidebarState &&
-              session.hasWorkflowError === hasWorkflowError)
-          ) {
-            return session;
-          }
-          changed = true;
-          return {
-            ...session,
-            hasRunningWorkflow,
-            workflowSidebarState,
-            hasWorkflowError,
-          };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [syncSessionWorkflowErrorIndicator],
-  );
-  const setSessionWorkflowStatusIndicators = useCallback(
-    (
-      sessionId: string,
-      status: {
-        sidebar_workflow_state?: WorkflowSidebarState | null;
-        has_running_workflow: boolean;
-        pending_workflow_input_id?: string | null;
-        pending_workflow_review_id?: string | null;
-      },
-    ) => {
-      if (!sessionId) return;
-      const workflowSidebarState = resolveWorkflowSidebarState(status);
-      const hasRunningWorkflow = isWorkflowSidebarRunning(workflowSidebarState);
-      const pendingWorkflowInputId = status.pending_workflow_input_id ?? null;
-      const pendingWorkflowReviewId = status.pending_workflow_review_id ?? null;
-      const hasPendingWorkflowInput = syncSessionWorkflowInputIndicator(
-        sessionId,
-        pendingWorkflowInputId,
-      );
-      const hasPendingWorkflowReview = Boolean(pendingWorkflowReviewId);
-      const hasWorkflowError = syncSessionWorkflowErrorIndicator(
-        sessionId,
-        workflowSidebarState,
-      );
-      setSessionsAsync((prev) => {
-        let changed = false;
-        const data = prev.data.map((session) => {
-          if (
-            session.id !== sessionId ||
-            (session.hasRunningWorkflow === hasRunningWorkflow &&
-              session.workflowSidebarState === workflowSidebarState &&
-              session.pendingWorkflowInputId === pendingWorkflowInputId &&
-              session.hasPendingWorkflowInput === hasPendingWorkflowInput &&
-              session.pendingWorkflowReviewId === pendingWorkflowReviewId &&
-              session.hasPendingWorkflowReview === hasPendingWorkflowReview &&
-              session.hasWorkflowError === hasWorkflowError)
-          ) {
-            return session;
-          }
-          changed = true;
-          return {
-            ...session,
-            hasRunningWorkflow,
-            workflowSidebarState,
-            pendingWorkflowInputId,
-            hasPendingWorkflowInput,
-            pendingWorkflowReviewId,
-            hasPendingWorkflowReview,
-            hasWorkflowError,
-          };
-        });
-        return changed ? { ...prev, data } : prev;
-      });
-    },
-    [syncSessionWorkflowErrorIndicator, syncSessionWorkflowInputIndicator],
-  );
-
-  const clearSessionScopedState = useCallback(() => {
-    activeSessionIdRef.current = '';
-    setActiveSessionId('');
-    setMessagesAsync(succeed([]));
-    setMembersAsync(succeed([]));
-    setMemberQueuesBySessionAgentId({});
-    setMainAgentName(null);
-  }, []);
-
-  const setSelectedProjectId = useCallback(
-    (id: string) => {
-      const previousProjectId = selectedProjectIdRef.current;
-      selectedProjectIdRef.current = id;
-      setSelectedProjectIdState(id);
-      try {
-        if (id) {
-          localStorage.setItem(SELECTED_PROJECT_ID_STORAGE_KEY, id);
-        } else {
-          localStorage.removeItem(SELECTED_PROJECT_ID_STORAGE_KEY);
-        }
-      } catch {}
-
-      if (previousProjectId !== id) {
-        inboxAutoReadProjectIdRef.current = null;
-        inboxInitialUnreadItemIdsRef.current = new Set();
-        autoMarkedInboxItemIdsRef.current = new Set();
-        setSessionsAsync(succeed([]));
-        setArchivedSessionsAsync(succeed([]));
-        clearSessionScopedState();
-      }
-    },
-    [clearSessionScopedState],
-  );
-
-  const syncSessionLeadAgent = useCallback(
-    async (sessionId: string, agentId: string | null): Promise<void> => {
-      if (!sessionId || !agentId) return;
-
-      const currentLeadAgentId =
-        sessionLeadAgentIdBySessionIdRef.current[sessionId] ?? null;
-      if (currentLeadAgentId === agentId) return;
-
-      sessionLeadAgentIdBySessionIdRef.current = {
-        ...sessionLeadAgentIdBySessionIdRef.current,
-        [sessionId]: agentId,
-      };
-
-      try {
-        const updatedSession = await chatSessionsApi.update(
-          sessionId,
-          chatSessionUpdatePayload({ lead_agent_id: agentId }),
-        );
-        sessionLeadAgentIdBySessionIdRef.current = {
-          ...sessionLeadAgentIdBySessionIdRef.current,
-          [updatedSession.id]: updatedSession.lead_agent_id,
-        };
-      } catch (err) {
-        sessionLeadAgentIdBySessionIdRef.current = {
-          ...sessionLeadAgentIdBySessionIdRef.current,
-          [sessionId]: currentLeadAgentId,
-        };
-        console.warn('Failed to sync workflow lead agent', err);
-      }
-    },
-    [],
-  );
-
-  const ensureWorkflowRouteToMainAgent = useCallback(async (): Promise<void> => {
-    const sid = activeSessionIdRef.current;
-    const agentId = workflowRouteAgentIdRef.current;
-    await syncSessionLeadAgent(sid, agentId);
-  }, [syncSessionLeadAgent]);
-
-  const setSessionChatInputMode = useCallback(
-    (sessionId: string, mode: ChatInputMode) => {
-      if (!sessionId) return;
-      setChatInputModeBySessionId((prev) => ({
-        ...prev,
-        [sessionId]: mode,
-      }));
-    },
-    [],
-  );
-
-  const setChatInputMode = useCallback(
-    (mode?: ChatInputMode) => {
-      const sid = activeSessionIdRef.current;
-      if (!sid) return;
-
-      const previousMode =
-        chatInputModeBySessionId[sid] ?? DEFAULT_CHAT_INPUT_MODE;
-      const nextMode =
-        mode ?? (previousMode === 'workflow' ? 'free' : 'workflow');
-
-      setChatInputModeBySessionId((prev) => ({
-        ...prev,
-        [sid]: nextMode,
-      }));
-      if (nextMode === 'workflow') {
-        void ensureWorkflowRouteToMainAgent();
-      }
-
-      if (sessionsAsync.source !== 'api') return;
-
-      chatSessionsApi
-        .update(sid, {
-          ...chatSessionUpdatePayload({
-            chat_input_mode: toSessionChatInputMode(nextMode),
-          }),
-        })
-        .then((updatedSession) => {
-          setChatInputModeBySessionId((prev) => ({
-            ...prev,
-            [updatedSession.id]: resolveChatInputMode(
-              updatedSession.chat_input_mode,
-            ),
-          }));
-        })
-        .catch((err) => {
-          setChatInputModeBySessionId((prev) => ({
-            ...prev,
-            [sid]: previousMode,
-          }));
-          showToast(
-            err instanceof Error
-              ? `Mode switch failed: ${err.message}`
-              : 'Mode switch failed.',
-            'error',
-          );
-        });
-    },
-    [chatInputModeBySessionId, ensureWorkflowRouteToMainAgent, sessionsAsync.source],
-  );
+    chatInputMode,
+    showToast,
+    persistUiPreference,
+    setTheme,
+    setLocale,
+    setChatMessageFontSize,
+    makeListSetter,
+    setSessions,
+    setMembers,
+    setProviders,
+    setSessionRunningIndicator,
+    syncProcessingQueuePlaceholders,
+    applyChatRuntimeSnapshot,
+    reconcileStartingPlaceholders,
+    setSessionWorkflowRunningIndicator,
+    setSessionWorkflowStatusIndicators,
+    clearSessionScopedState,
+    setSelectedProjectId,
+    syncSessionLeadAgent,
+    ensureWorkflowRouteToMainAgent,
+    setSessionChatInputMode,
+    setChatInputMode,
+  } = workspaceState;
 
   const applyMockBootstrap = useCallback(
     (bootstrap: WorkspaceBootstrapMock) => {
@@ -3867,495 +1378,19 @@ export const WorkspaceProvider: React.FC<{ children: React.ReactNode }> = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const mapBackendChatMessage = useCallback(
-    (message: BackendChatMessage): Message =>
-      mapMessage(message, {
-        agentNamesById: agentNamesByIdRef.current,
-        agentModelsById: agentModelsByIdRef.current,
-      }),
-    [],
-  );
-
-  const insertQueuedBackendUserMessage = useCallback(
-    (sid: string, runId: string, message: Message) => {
-      setAllMessages((prev) => {
-        const current = filterMessagesForSession(sid, prev[sid] ?? []);
-        const sourceClientMessageId = message.isUser
-          ? userMessageClientId(message)
-          : undefined;
-        const withoutExistingSourceMessage = current.filter((candidate) =>
-          message.isUser
-            ? !matchesUserMessageIdentity(
-                candidate,
-                message.id,
-                sourceClientMessageId,
-              )
-            : candidate.id !== message.id,
-        );
-
-        const runIndex = withoutExistingSourceMessage.findIndex(
-          (candidate) => candidate.isAgentRunning && candidate.runId === runId,
-        );
-        const next = [...withoutExistingSourceMessage];
-        next.splice(runIndex >= 0 ? runIndex : next.length, 0, message);
-        return { ...prev, [sid]: resolveMessageReferences(next) };
-      });
-    },
-    [],
-  );
-
-  const ensureQueuedRunSourceMessage = useCallback(
-    async (
-      event: Extract<ChatStreamEvent, { type: 'agent_run_started' }>,
-    ): Promise<void> => {
-      try {
-        const backendMessage = await chatMessagesApi.get(
-          event.source_message_id,
-        );
-        insertQueuedBackendUserMessage(
-          event.session_id,
-          event.run_id,
-          mapBackendChatMessage(backendMessage),
-        );
-      } catch {
-        // Source-message hydration is best-effort; the running placeholder still shows.
-      }
-    },
-    [insertQueuedBackendUserMessage, mapBackendChatMessage],
-  );
-
-  const upsertStreamedMessage = useCallback(
-    (sid: string, incoming: Message) => {
-      setAllMessages((prev) => {
-        const current = filterMessagesForSession(sid, prev[sid] ?? []);
-        let carriedSessionAgentId = incoming.sessionAgentId;
-        let carriedSourceMessageId = incoming.sourceMessageId;
-        let carriedClientMessageId = incoming.clientMessageId;
-        const matchingPlaceholderIndexes = new Set(
-          findRunningPlaceholderIndexesForIncoming(current, incoming),
-        );
-        let replacementIndex: number | null = null;
-        const withoutPlaceholder = current.filter((message, index) => {
-          if (matchingPlaceholderIndexes.has(index)) {
-            replacementIndex =
-              replacementIndex === null
-                ? index
-                : Math.min(replacementIndex, index);
-            carriedSessionAgentId =
-              carriedSessionAgentId ?? message.sessionAgentId;
-            carriedSourceMessageId =
-              carriedSourceMessageId ?? message.sourceMessageId;
-            carriedClientMessageId =
-              carriedClientMessageId ?? message.clientMessageId;
-            return false;
-          }
-          return true;
-        });
-        const nextMessage: Message = {
-          ...incoming,
-          sessionAgentId: carriedSessionAgentId,
-          sourceMessageId: carriedSourceMessageId,
-          clientMessageId: carriedClientMessageId,
-          isAgentRunning: undefined,
-          isThinking: undefined,
-        };
-        if (!nextMessage.isUser && nextMessage.sessionAgentId) {
-          optimisticallyStoppedSessionAgentIdsRef.current.delete(
-            nextMessage.sessionAgentId,
-          );
-        }
-        const nextClientMessageId = userMessageClientId(nextMessage);
-        const existingIndex = withoutPlaceholder.findIndex((message) => {
-          if (message.id === nextMessage.id) return true;
-          return (
-            nextMessage.isUser &&
-            nextClientMessageId !== undefined &&
-            userMessageClientId(message) === nextClientMessageId
-          );
-        });
-        const next =
-          existingIndex >= 0
-            ? withoutPlaceholder.map((message, index) =>
-                index === existingIndex ? nextMessage : message,
-              )
-            : (() => {
-                const inserted = [...withoutPlaceholder];
-                inserted.splice(
-                  replacementIndex === null
-                    ? inserted.length
-                    : Math.min(replacementIndex, inserted.length),
-                  0,
-                  nextMessage,
-                );
-                return inserted;
-              })();
-        const correlatedNext =
-          nextMessage.isUser && nextClientMessageId
-            ? next.map((message) =>
-                isPendingAgentPlaceholder(message) &&
-                message.clientMessageId === nextClientMessageId
-                  ? { ...message, sourceMessageId: nextMessage.id }
-                  : message,
-              )
-            : next;
-        return {
-          ...prev,
-          [sid]: resolveMessageReferences(
-            orderMessagesForConversation(correlatedNext),
-          ),
-        };
-      });
-    },
-    [],
-  );
-
-  const insertRunningPlaceholder = useCallback(
-    (event: Extract<ChatStreamEvent, { type: 'agent_run_started' }>) => {
-      // A new run for this agent supersedes any optimistic-stop suppression.
-      optimisticallyStoppedSessionAgentIdsRef.current.delete(
-        event.session_agent_id,
-      );
-      setSessionRunningIndicator(event.session_id, true);
-      void ensureQueuedRunSourceMessage(event);
-      setActiveRunsByRunId((prev) => {
-        const displayName = event.agent_name.startsWith('@')
-          ? event.agent_name
-          : `@${event.agent_name}`;
-        const nextRun: RuntimeActiveRun = {
-          run_id: event.run_id,
-          session_id: event.session_id,
-          session_agent_id: event.session_agent_id,
-          agent_id: event.agent_id,
-          agent_name: event.agent_name,
-          display_name: displayName,
-          avatar: monogramFromName(event.agent_name),
-          model: event.model ?? agentModelsByIdRef.current[event.agent_id] ?? null,
-          status: 'running',
-          source_message_id: event.source_message_id,
-          client_message_id: event.client_message_id ?? null,
-          created_at: event.started_at ?? new Date().toISOString(),
-        };
-        const next = { ...prev };
-        for (const [runId, run] of Object.entries(next)) {
-          if (
-            run.session_agent_id === event.session_agent_id &&
-            runId !== event.run_id
-          ) {
-            delete next[runId];
-          }
-        }
-        next[event.run_id] = nextRun;
-        return next;
-      });
-    },
-    [ensureQueuedRunSourceMessage, setSessionRunningIndicator],
-  );
-
-  const handleWorkflowRuntimeLine = useCallback(
-    (event: Extract<ChatStreamEvent, { type: 'workflow_runtime_line' }>) => {
-      setWorkflowRuntimeLinesByExecution((prev) => {
-        const executionLines = prev[event.execution_id] ?? [];
-        if (executionLines.some((line) => line.id === event.line_id)) {
-          return prev;
-        }
-
-        return {
-          ...prev,
-          [event.execution_id]: [
-            ...executionLines,
-            {
-              id: event.line_id,
-              executionId: event.execution_id,
-              workflowAgentSessionId: event.workflow_agent_session_id,
-              stepId: event.step_id,
-              stepKey: event.step_key,
-              agentId: event.agent_id,
-              agentName: event.agent_name,
-              streamType: event.stream_type,
-              content: event.content,
-              createdAt: event.created_at,
-            },
-          ],
-        };
-      });
-    },
-    [],
-  );
-
-  // When the active session changes, re-fetch its scoped data.
-  useEffect(() => {
-    if (!activeSessionId) return;
-    void refreshMessages();
-    void refreshMembers();
-    void refreshMemberQueues();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeSessionId]);
-
-  useEffect(() => {
-    if (!activeSessionId || sessionsAsync.source !== 'api') return;
-
-    const sid = activeSessionId;
-    let socket: WebSocket | null = null;
-    let reconnectTimer: ReturnType<typeof setTimeout> | null = null;
-    let reconnectAttempt = 0;
-    let hasConnectedOnce = false;
-    let disposed = false;
-
-    const handleMessage = (event: MessageEvent) => {
-      let parsed: ChatStreamEvent;
-      try {
-        parsed = JSON.parse(event.data) as ChatStreamEvent;
-      } catch {
-        return;
-      }
-
-      if (parsed.type === 'agent_run_started' && parsed.session_id === sid) {
-        insertRunningPlaceholder(parsed);
-        return;
-      }
-
-      if (
-        parsed.type === 'agent_activity_updated' &&
-        parsed.session_id === sid
-      ) {
-        runActivityStore.notifyUpdated(
-          parsed.run_id,
-          parsed.latest_sequence,
-        );
-        return;
-      }
-
-      if (
-        parsed.type === 'workflow_runtime_line' &&
-        parsed.session_id === sid
-      ) {
-        setSessionWorkflowRunningIndicator(sid, true);
-        handleWorkflowRuntimeLine(parsed);
-        return;
-      }
-
-      if (
-        parsed.type === 'workflow_execution_updated' &&
-        parsed.session_id === sid
-      ) {
-        void refreshSessionRunningIndicators(sid);
-        scheduleInboxRefresh();
-        return;
-      }
-
-      if (
-        parsed.type === 'file_change_refresh' &&
-        parsed.session_id === sid
-      ) {
-        const projectId = selectedProjectIdRef.current;
-        notifySourceControlRefreshRequested({
-          projectId,
-          sessionId: sid,
-        });
-        const workspacePath = activeWorkspacePathRef.current;
-        if (!projectId && workspacePath) {
-          void refreshWorkspaceChanges(sid, workspacePath, true);
-        }
-        return;
-      }
-
-      if (parsed.type === 'queue_updated' && parsed.session_id === sid) {
-        mergeMemberQueueSnapshot(parsed.queue);
-        return;
-      }
-
-      if (
-        (parsed.type === 'message_new' || parsed.type === 'message_updated') &&
-        parsed.message.session_id === sid
-      ) {
-        const tokenUsageSignature = tokenUsageNotificationSignature(
-          parsed.message,
-        );
-        if (
-          tokenUsageSignature &&
-          notifiedTokenUsageSignaturesRef.current[parsed.message.id] !==
-            tokenUsageSignature
-        ) {
-          notifiedTokenUsageSignaturesRef.current[parsed.message.id] =
-            tokenUsageSignature;
-          const projectId = selectedProjectIdRef.current;
-          if (projectId) {
-            notifyBuildStatsUsageUpdated(projectId);
-          }
-        }
-        const incomingMessage = mapBackendChatMessage(parsed.message);
-        upsertStreamedMessage(sid, incomingMessage);
-        if (incomingMessage.runId) {
-          setActiveRunsByRunId((prev) => {
-            if (!incomingMessage.runId || !prev[incomingMessage.runId]) {
-              return prev;
-            }
-            const next = { ...prev };
-            delete next[incomingMessage.runId];
-            return next;
-          });
-        }
-        scheduleInboxRefresh();
-        return;
-      }
-
-      if (parsed.type === 'agent_state') {
-        if (
-          parsed.run_id &&
-          !isRunningSessionAgentState(parsed.state)
-        ) {
-          runActivityStore.requestCompletion(parsed.run_id);
-        }
-        if (isRunningSessionAgentState(parsed.state)) {
-          setSessionRunningIndicator(sid, true);
-        } else {
-          setActiveRunsByRunId((prev) => {
-            const next = { ...prev };
-            let changed = false;
-            for (const [runId, run] of Object.entries(next)) {
-              if (
-                run.session_agent_id === parsed.session_agent_id &&
-                (!parsed.run_id || runId === parsed.run_id)
-              ) {
-                delete next[runId];
-                changed = true;
-              }
-            }
-            const hasRemainingRunningAgent = Object.values(next).some(
-              (run) => run.session_id === sid,
-            );
-            setSessionRunningIndicator(sid, hasRemainingRunningAgent);
-            return changed ? next : prev;
-          });
-          void refreshSessionWorkflowStatus(sid);
-        }
-        void refreshMembers();
-        return;
-      }
-
-      if (parsed.type === 'mention_error' && parsed.session_id === sid) {
-        if (parsed.reason === 'member_not_found') {
-          showToast(
-            memberNotFoundToastMessage(locale, parsed.agent_name),
-            'error',
-          );
-        }
-        setAllMessages((prev) => {
-          const current = filterMessagesForSession(sid, prev[sid] ?? []);
-          if (current.length === 0) return prev;
-          const updated = current.filter(
-            (msg) =>
-              !(
-                isOptimisticPendingAgentPlaceholder(msg) &&
-                pendingPlaceholderMatches(msg, {
-                  clientMessageId: parsed.client_message_id,
-                  sourceMessageId: parsed.message_id,
-                  agentName: parsed.agent_name,
-                })
-              ),
-          );
-          if (updated.length === current.length) return prev;
-          return { ...prev, [sid]: updated };
-        });
-      }
-    };
-
-    // Open the stream and keep it alive across transient drops. The stream has
-    // no server-side replay, so on every *re*connect we re-hydrate the session
-    // via REST to recover any persisted messages emitted while we were down.
-    const connect = () => {
-      if (disposed) return;
-      const ws = new WebSocket(
-        chatStreamWebSocketUrl(chatSessionsApi.streamUrl(sid)),
-      );
-      socket = ws;
-      ws.onmessage = handleMessage;
-      ws.onopen = () => {
-        reconnectAttempt = 0;
-        if (hasConnectedOnce) {
-          void refreshMessages();
-          void refreshMembers();
-          void refreshMemberQueues();
-          const projectId = selectedProjectIdRef.current;
-          if (projectId) {
-            notifySourceControlRefreshRequested({
-              projectId,
-              sessionId: sid,
-            });
-          }
-          const workspacePath = activeWorkspacePathRef.current;
-          if (!projectId && workspacePath) {
-            void refreshWorkspaceChanges(sid, workspacePath, true);
-          }
-        }
-        hasConnectedOnce = true;
-      };
-      ws.onclose = () => {
-        // Ignore the close of a superseded socket or one closed by cleanup.
-        if (disposed || socket !== ws) return;
-        const delay = Math.min(
-          CHAT_STREAM_RECONNECT_BASE_DELAY_MS * 2 ** reconnectAttempt,
-          CHAT_STREAM_RECONNECT_MAX_DELAY_MS,
-        );
-        reconnectAttempt += 1;
-        reconnectTimer = setTimeout(connect, delay);
-      };
-      // Let onclose drive the reconnect; just tear the socket down on error.
-      ws.onerror = () => {
-        ws.close();
-      };
-    };
-
-    connect();
-
-    return () => {
-      disposed = true;
-      if (reconnectTimer) {
-        clearTimeout(reconnectTimer);
-        reconnectTimer = null;
-      }
-      socket?.close();
-    };
-  }, [
-    activeSessionId,
-    handleWorkflowRuntimeLine,
-    insertRunningPlaceholder,
-    locale,
-    mapBackendChatMessage,
-    mergeMemberQueueSnapshot,
-    refreshMessages,
-    refreshMemberQueues,
-    refreshSessionRunningIndicators,
-    refreshSessionWorkflowStatus,
-    refreshWorkspaceChanges,
-    refreshMembers,
-    runActivityStore,
-    scheduleInboxRefresh,
-    setSessionRunningIndicator,
-    setSessionWorkflowRunningIndicator,
-    sessionsAsync.source,
-    upsertStreamedMessage,
-  ]);
-
-  useEffect(() => {
-    const syncVisibleRuns = () => {
-      if (document.visibilityState !== 'visible' || !activeSessionId) return;
-      runActivityStore.syncRuns(
-        Object.values(activeRunsByRunId)
-          .filter((run) => run.session_id === activeSessionId)
-          .map((run) => run.run_id),
-      );
-    };
-    document.addEventListener('visibilitychange', syncVisibleRuns);
-    return () => {
-      document.removeEventListener('visibilitychange', syncVisibleRuns);
-    };
-  }, [activeRunsByRunId, activeSessionId, runActivityStore]);
-
-  useEffect(() => {
-    if (!initialRefreshCompletedRef.current) return;
-    void refreshSessions();
-  }, [refreshSessions, selectedProjectId]);
+  const { mapBackendChatMessage, upsertStreamedMessage } =
+    useWorkspaceChatRuntime({
+      ...workspaceState,
+      mergeMemberQueueSnapshot,
+      refreshMemberQueues,
+      refreshMembers,
+      refreshMessages,
+      refreshSessionRunningIndicators,
+      refreshSessionWorkflowStatus,
+      refreshSessions,
+      refreshWorkspaceChanges,
+      scheduleInboxRefresh,
+    });
 
   // ---------------------------------------------------------------------------
   // i18n
